Move prompt editor and codegen into their own modules
