+++ crates/assistant2/src/inline_assistant.rs
@@ -1,74 +1,46 @@
-use crate::context::attach_context_to_message;
-use crate::context_picker::ContextPicker;
+use crate::buffer_codegen::{BufferCodegen, CodegenAlternative};
 use crate::context_store::ContextStore;
-use crate::context_strip::ContextStrip;
 use crate::inline_prompt_editor::{
-    render_cancel_button, CodegenStatus, PromptEditorEvent, PromptMode,
+    CodegenEvent, CodegenStatus, InlineAssistId, PromptEditor, PromptEditorEvent, PromptEditorMode,
 };
 use crate::thread_store::ThreadStore;
+use crate::AssistantPanel;
 use crate::{
-    assistant_settings::AssistantSettings,
-    prompts::PromptBuilder,
-    streaming_diff::{CharOperation, LineDiff, LineOperation, StreamingDiff},
+    assistant_settings::AssistantSettings, prompts::PromptBuilder,
     terminal_inline_assistant::TerminalInlineAssistant,
-    CycleNextInlineAssist, CyclePreviousInlineAssist,
 };
-use crate::{AssistantPanel, ToggleContextPicker};
 use anyhow::{Context as _, Result};
-use client::{telemetry::Telemetry, ErrorExt};
+use client::telemetry::Telemetry;
 use collections::{hash_map, HashMap, HashSet, VecDeque};
 use editor::{
-    actions::{MoveDown, MoveUp, SelectAll},
+    actions::SelectAll,
     display_map::{
         BlockContext, BlockPlacement, BlockProperties, BlockStyle, CustomBlockId, RenderBlock,
         ToDisplayPoint,
     },
-    Anchor, AnchorRangeExt, CodeActionProvider, Editor, EditorElement, EditorEvent, EditorMode,
-    EditorStyle, ExcerptId, ExcerptRange, GutterDimensions, MultiBuffer, MultiBufferSnapshot,
-    ToOffset as _, ToPoint,
+    Anchor, AnchorRangeExt, CodeActionProvider, Editor, EditorEvent, ExcerptId, ExcerptRange,
+    GutterDimensions, MultiBuffer, MultiBufferSnapshot, ToOffset as _, ToPoint,
 };
-use feature_flags::{FeatureFlagAppExt as _, ZedPro};
 use fs::Fs;
-use futures::{channel::mpsc, future::LocalBoxFuture, join, SinkExt, Stream, StreamExt};
+use util::ResultExt;
+
 use gpui::{
-    anchored, deferred, point, AnyElement, AppContext, ClickEvent, CursorStyle, EventEmitter,
-    FocusHandle, FocusableView, FontWeight, Global, HighlightStyle, Model, ModelContext,
-    Subscription, Task, TextStyle, UpdateGlobal, View, ViewContext, WeakModel, WeakView,
-    WindowContext,
+    point, AppContext, FocusableView, Global, HighlightStyle, Model, Subscription, Task,
+    UpdateGlobal, View, ViewContext, WeakModel, WeakView, WindowContext,
 };
-use language::{Buffer, IndentKind, Point, Selection, TransactionId};
-use language_model::{
-    LanguageModel, LanguageModelRegistry, LanguageModelRequest, LanguageModelRequestMessage,
-    LanguageModelTextStream, Role,
-};
-use language_model_selector::{LanguageModelSelector, LanguageModelSelectorPopoverMenu};
+use language::{Buffer, Point, Selection, TransactionId};
+use language_model::LanguageModelRegistry;
 use language_models::report_assistant_event;
 use multi_buffer::MultiBufferRow;
 use parking_lot::Mutex;
 use project::{CodeAction, ProjectTransaction};
-use rope::Rope;
-use settings::{update_settings_file, Settings, SettingsStore};
-use smol::future::FutureExt;
-use std::{
-    cmp,
-    future::Future,
-    iter, mem,
-    ops::{Range, RangeInclusive},
-    pin::Pin,
-    rc::Rc,
-    sync::Arc,
-    task::{self, Poll},
-    time::Instant,
-};
+use settings::{Settings, SettingsStore};
+use std::{cmp, mem, ops::Range, rc::Rc, sync::Arc};
 use telemetry_events::{AssistantEvent, AssistantKind, AssistantPhase};
-use terminal::Terminal;
 use terminal_view::{terminal_panel::TerminalPanel, TerminalView};
 use text::{OffsetRangeExt, ToPoint as _};
-use theme::ThemeSettings;
-use ui::{
-    prelude::*, CheckboxWithLabel, IconButtonShape, KeyBinding, Popover, PopoverMenuHandle, Tooltip,
-};
-use util::{RangeExt, ResultExt};
+use ui::prelude::*;
+use util::RangeExt;
 use workspace::{dock::Panel, ShowConfiguration};
 use workspace::{notifications::NotificationId, ItemHandle, Toast, Workspace};
 
@@ -424,7 +396,7 @@ impl InlineAssistant {
         let mut assist_group = InlineAssistGroup::new();
         for (assist_id, range, prompt_editor, prompt_block_id, end_block_id) in assists {
             let codegen = match &prompt_editor.read(cx).mode {
-                PromptEditorMode::Buffer { codegen } => codegen.clone(),
+                PromptEditorMode::Buffer { codegen, .. } => codegen.clone(),
                 PromptEditorMode::Terminal { .. } => unreachable!("Unexpected terminal mode"),
             };
 
@@ -1459,21 +1431,19 @@ impl InlineAssistGroup {
 
 fn build_assist_editor_renderer(editor: &View<PromptEditor>) -> RenderBlock {
     let editor = editor.clone();
+
     Arc::new(move |cx: &mut BlockContext| {
-        *editor.read(cx).gutter_dimensions.lock() = *cx.gutter_dimensions;
+        let gutter_dimensions = match &editor.read(cx).mode {
+            PromptEditorMode::Terminal { .. } => unreachable!("Unexpected terminal mode"),
+            PromptEditorMode::Buffer {
+                gutter_dimensions,
+                codegen: _,
+            } => gutter_dimensions,
+        };
+
+        *gutter_dimensions.lock() = *cx.gutter_dimensions;
         editor.clone().into_any_element()
     })
-}
-
-#[derive(Copy, Clone, Default, Debug, PartialEq, Eq, Hash)]
-pub struct InlineAssistId(usize);
-
-impl InlineAssistId {
-    fn post_inc(&mut self) -> InlineAssistId {
-        let id = *self;
-        self.0 += 1;
-        id
-    }
 }
 
 #[derive(Copy, Clone, Default, Debug, PartialEq, Eq, Hash)]
@@ -1485,732 +1455,6 @@ impl InlineAssistGroupId {
         self.0 += 1;
         id
     }
-}
-
-struct PromptEditor {
-    id: InlineAssistId,
-    editor: View<Editor>,
-    context_strip: View<ContextStrip>,
-    context_picker_menu_handle: PopoverMenuHandle<ContextPicker>,
-    language_model_selector: View<LanguageModelSelector>,
-    edited_since_done: bool,
-    gutter_dimensions: Arc<Mutex<GutterDimensions>>,
-    prompt_history: VecDeque<String>,
-    prompt_history_ix: Option<usize>,
-    pending_prompt: String,
-    _codegen_subscription: Subscription,
-    editor_subscriptions: Vec<Subscription>,
-    show_rate_limit_notice: bool,
-    mode: PromptEditorMode,
-}
-
-impl EventEmitter<PromptEditorEvent> for PromptEditor {}
-
-impl Render for PromptEditor {
-    fn render(&mut self, cx: &mut ViewContext<Self>) -> impl IntoElement {
-        let gutter_dimensions = *self.gutter_dimensions.lock();
-        let mut buttons = Vec::new();
-
-        // TODO (agus): remove prompt mode and move buttons view here
-        let prompt_mode = match &self.mode {
-            PromptEditorMode::Buffer { codegen } => {
-                let codegen = codegen.read(cx);
-
-                if codegen.alternative_count(cx) > 1 {
-                    buttons.push(self.render_cycle_controls(cx));
-                }
-
-                if codegen.is_insertion {
-                    PromptMode::Generate {
-                        supports_execute: false,
-                    }
-                } else {
-                    PromptMode::Transform
-                }
-            }
-            PromptEditorMode::Terminal { codegen: _ } => {
-                // todo(terminal-prompt-cycle)
-                PromptMode::Generate {
-                    supports_execute: true,
-                }
-            }
-        };
-
-        buttons.extend(render_cancel_button(
-            self.codegen_status(cx).into(),
-            self.edited_since_done,
-            prompt_mode,
-            cx,
-        ));
-
-        v_flex()
-            .border_y_1()
-            .border_color(cx.theme().status().info_border)
-            .size_full()
-            .py(cx.line_height() / 2.5)
-            .child(
-                h_flex()
-                    .key_context("PromptEditor")
-                    .bg(cx.theme().colors().editor_background)
-                    .block_mouse_down()
-                    .cursor(CursorStyle::Arrow)
-                    .on_action(cx.listener(Self::toggle_context_picker))
-                    .on_action(cx.listener(Self::confirm))
-                    .on_action(cx.listener(Self::cancel))
-                    .on_action(cx.listener(Self::move_up))
-                    .on_action(cx.listener(Self::move_down))
-                    .capture_action(cx.listener(Self::cycle_prev))
-                    .capture_action(cx.listener(Self::cycle_next))
-                    .child(
-                        h_flex()
-                            .w(gutter_dimensions.full_width() + (gutter_dimensions.margin / 2.0))
-                            .justify_center()
-                            .gap_2()
-                            .child(LanguageModelSelectorPopoverMenu::new(
-                                self.language_model_selector.clone(),
-                                IconButton::new("context", IconName::SettingsAlt)
-                                    .shape(IconButtonShape::Square)
-                                    .icon_size(IconSize::Small)
-                                    .icon_color(Color::Muted)
-                                    .tooltip(move |cx| {
-                                        Tooltip::with_meta(
-                                            format!(
-                                                "Using {}",
-                                                LanguageModelRegistry::read_global(cx)
-                                                    .active_model()
-                                                    .map(|model| model.name().0)
-                                                    .unwrap_or_else(|| "No model selected".into()),
-                                            ),
-                                            None,
-                                            "Change Model",
-                                            cx,
-                                        )
-                                    }),
-                            ))
-                            .map(|el| {
-                                let CodegenStatus::Error(error) = self.codegen_status(cx) else {
-                                    return el;
-                                };
-
-                                let error_message = SharedString::from(error.to_string());
-                                if error.error_code() == proto::ErrorCode::RateLimitExceeded
-                                    && cx.has_flag::<ZedPro>()
-                                {
-                                    el.child(
-                                        v_flex()
-                                            .child(
-                                                IconButton::new(
-                                                    "rate-limit-error",
-                                                    IconName::XCircle,
-                                                )
-                                                .toggle_state(self.show_rate_limit_notice)
-                                                .shape(IconButtonShape::Square)
-                                                .icon_size(IconSize::Small)
-                                                .on_click(
-                                                    cx.listener(Self::toggle_rate_limit_notice),
-                                                ),
-                                            )
-                                            .children(self.show_rate_limit_notice.then(|| {
-                                                deferred(
-                                                    anchored()
-                                                        .position_mode(
-                                                            gpui::AnchoredPositionMode::Local,
-                                                        )
-                                                        .position(point(px(0.), px(24.)))
-                                                        .anchor(gpui::Corner::TopLeft)
-                                                        .child(self.render_rate_limit_notice(cx)),
-                                                )
-                                            })),
-                                    )
-                                } else {
-                                    el.child(
-                                        div()
-                                            .id("error")
-                                            .tooltip(move |cx| {
-                                                Tooltip::text(error_message.clone(), cx)
-                                            })
-                                            .child(
-                                                Icon::new(IconName::XCircle)
-                                                    .size(IconSize::Small)
-                                                    .color(Color::Error),
-                                            ),
-                                    )
-                                }
-                            }),
-                    )
-                    .child(div().flex_1().child(self.render_editor(cx)))
-                    .child(h_flex().gap_2().pr_6().children(buttons)),
-            )
-            .child(
-                h_flex()
-                    .child(
-                        h_flex()
-                            .w(gutter_dimensions.full_width() + (gutter_dimensions.margin / 2.0))
-                            .justify_center()
-                            .gap_2(),
-                    )
-                    .child(self.context_strip.clone()),
-            )
-    }
-}
-
-impl FocusableView for PromptEditor {
-    fn focus_handle(&self, cx: &AppContext) -> FocusHandle {
-        self.editor.focus_handle(cx)
-    }
-}
-
-impl PromptEditor {
-    const MAX_LINES: u8 = 8;
-
-    #[allow(clippy::too_many_arguments)]
-    fn new_buffer(
-        id: InlineAssistId,
-        gutter_dimensions: Arc<Mutex<GutterDimensions>>,
-        prompt_history: VecDeque<String>,
-        prompt_buffer: Model<MultiBuffer>,
-        codegen: Model<BufferCodegen>,
-        fs: Arc<dyn Fs>,
-        context_store: Model<ContextStore>,
-        workspace: WeakView<Workspace>,
-        thread_store: Option<WeakModel<ThreadStore>>,
-        cx: &mut ViewContext<Self>,
-    ) -> Self {
-        let codegen_subscription = cx.observe(&codegen, Self::handle_codegen_changed);
-        let mode = PromptEditorMode::Buffer { codegen };
-
-        let prompt_editor = cx.new_view(|cx| {
-            let mut editor = Editor::new(
-                EditorMode::AutoHeight {
-                    max_lines: Self::MAX_LINES as usize,
-                },
-                prompt_buffer,
-                None,
-                false,
-                cx,
-            );
-            editor.set_soft_wrap_mode(language::language_settings::SoftWrap::EditorWidth, cx);
-            // Since the prompt editors for all inline assistants are linked,
-            // always show the cursor (even when it isn't focused) because
-            // typing in one will make what you typed appear in all of them.
-            editor.set_show_cursor_when_unfocused(true, cx);
-            editor.set_placeholder_text(Self::placeholder_text(&mode, cx), cx);
-            editor
-        });
-        let context_picker_menu_handle = PopoverMenuHandle::default();
-
-        let mut this = Self {
-            id,
-            editor: prompt_editor.clone(),
-            context_strip: cx.new_view(|cx| {
-                ContextStrip::new(
-                    context_store,
-                    workspace.clone(),
-                    thread_store.clone(),
-                    prompt_editor.focus_handle(cx),
-                    context_picker_menu_handle.clone(),
-                    cx,
-                )
-            }),
-            context_picker_menu_handle,
-            language_model_selector: cx.new_view(|cx| {
-                let fs = fs.clone();
-                LanguageModelSelector::new(
-                    move |model, cx| {
-                        update_settings_file::<AssistantSettings>(
-                            fs.clone(),
-                            cx,
-                            move |settings, _| settings.set_model(model.clone()),
-                        );
-                    },
-                    cx,
-                )
-            }),
-            edited_since_done: false,
-            gutter_dimensions,
-            prompt_history,
-            prompt_history_ix: None,
-            pending_prompt: String::new(),
-            _codegen_subscription: codegen_subscription,
-            editor_subscriptions: Vec::new(),
-            show_rate_limit_notice: false,
-            mode,
-        };
-        this.subscribe_to_editor(cx);
-        this
-    }
-
-    fn codegen_status<'a>(&'a self, cx: &'a AppContext) -> &'a CodegenStatus {
-        match &self.mode {
-            PromptEditorMode::Buffer { codegen } => codegen.read(cx).status(cx),
-            PromptEditorMode::Terminal { codegen } => &codegen.read(cx).status,
-        }
-    }
-
-    fn subscribe_to_editor(&mut self, cx: &mut ViewContext<Self>) {
-        self.editor_subscriptions.clear();
-        self.editor_subscriptions
-            .push(cx.subscribe(&self.editor, Self::handle_prompt_editor_events));
-    }
-
-    fn set_show_cursor_when_unfocused(
-        &mut self,
-        show_cursor_when_unfocused: bool,
-        cx: &mut ViewContext<Self>,
-    ) {
-        self.editor.update(cx, |editor, cx| {
-            editor.set_show_cursor_when_unfocused(show_cursor_when_unfocused, cx)
-        });
-    }
-
-    fn unlink(&mut self, cx: &mut ViewContext<Self>) {
-        let prompt = self.prompt(cx);
-        let focus = self.editor.focus_handle(cx).contains_focused(cx);
-        self.editor = cx.new_view(|cx| {
-            let mut editor = Editor::auto_height(Self::MAX_LINES as usize, cx);
-            editor.set_soft_wrap_mode(language::language_settings::SoftWrap::EditorWidth, cx);
-            editor.set_placeholder_text(Self::placeholder_text(&self.mode, cx), cx);
-            editor.set_placeholder_text("Add a prompt…", cx);
-            editor.set_text(prompt, cx);
-            if focus {
-                editor.focus(cx);
-            }
-            editor
-        });
-        self.subscribe_to_editor(cx);
-    }
-
-    pub fn placeholder_text(mode: &PromptEditorMode, cx: &WindowContext) -> String {
-        let action = match mode {
-            PromptEditorMode::Buffer { codegen } => {
-                if codegen.read(cx).is_insertion {
-                    "Generate"
-                } else {
-                    "Transform"
-                }
-            }
-            PromptEditorMode::Terminal { .. } => "Generate",
-        };
-
-        let assistant_panel_keybinding = ui::text_for_action(&crate::ToggleFocus, cx)
-            .map(|keybinding| format!("{keybinding} to chat ― "))
-            .unwrap_or_default();
-
-        format!("{action}… ({assistant_panel_keybinding}↓↑ for history)")
-    }
-
-    fn prompt(&self, cx: &AppContext) -> String {
-        self.editor.read(cx).text(cx)
-    }
-
-    fn toggle_rate_limit_notice(&mut self, _: &ClickEvent, cx: &mut ViewContext<Self>) {
-        self.show_rate_limit_notice = !self.show_rate_limit_notice;
-        if self.show_rate_limit_notice {
-            cx.focus_view(&self.editor);
-        }
-        cx.notify();
-    }
-
-    fn handle_prompt_editor_events(
-        &mut self,
-        _: View<Editor>,
-        event: &EditorEvent,
-        cx: &mut ViewContext<Self>,
-    ) {
-        match event {
-            EditorEvent::Edited { .. } => {
-                if let Some(workspace) = cx.window_handle().downcast::<Workspace>() {
-                    workspace
-                        .update(cx, |workspace, cx| {
-                            let is_via_ssh = workspace
-                                .project()
-                                .update(cx, |project, _| project.is_via_ssh());
-
-                            workspace
-                                .client()
-                                .telemetry()
-                                .log_edit_event("inline assist", is_via_ssh);
-                        })
-                        .log_err();
-                }
-                let prompt = self.editor.read(cx).text(cx);
-                if self
-                    .prompt_history_ix
-                    .map_or(true, |ix| self.prompt_history[ix] != prompt)
-                {
-                    self.prompt_history_ix.take();
-                    self.pending_prompt = prompt;
-                }
-
-                self.edited_since_done = true;
-                cx.notify();
-            }
-            EditorEvent::Blurred => {
-                if self.show_rate_limit_notice {
-                    self.show_rate_limit_notice = false;
-                    cx.notify();
-                }
-            }
-            _ => {}
-        }
-    }
-
-    fn handle_codegen_changed(&mut self, _: Model<BufferCodegen>, cx: &mut ViewContext<Self>) {
-        match self.codegen_status(cx) {
-            CodegenStatus::Idle => {
-                self.editor
-                    .update(cx, |editor, _| editor.set_read_only(false));
-            }
-            CodegenStatus::Pending => {
-                self.editor
-                    .update(cx, |editor, _| editor.set_read_only(true));
-            }
-            CodegenStatus::Done => {
-                self.edited_since_done = false;
-                self.editor
-                    .update(cx, |editor, _| editor.set_read_only(false));
-            }
-            CodegenStatus::Error(error) => {
-                if cx.has_flag::<ZedPro>()
-                    && error.error_code() == proto::ErrorCode::RateLimitExceeded
-                    && !dismissed_rate_limit_notice()
-                {
-                    self.show_rate_limit_notice = true;
-                    cx.notify();
-                }
-
-                self.edited_since_done = false;
-                self.editor
-                    .update(cx, |editor, _| editor.set_read_only(false));
-            }
-        }
-    }
-
-    fn toggle_context_picker(&mut self, _: &ToggleContextPicker, cx: &mut ViewContext<Self>) {
-        self.context_picker_menu_handle.toggle(cx);
-    }
-
-    fn cancel(&mut self, _: &editor::actions::Cancel, cx: &mut ViewContext<Self>) {
-        match self.codegen_status(cx) {
-            CodegenStatus::Idle | CodegenStatus::Done | CodegenStatus::Error(_) => {
-                cx.emit(PromptEditorEvent::CancelRequested);
-            }
-            CodegenStatus::Pending => {
-                cx.emit(PromptEditorEvent::StopRequested);
-            }
-        }
-    }
-
-    fn confirm(&mut self, _: &menu::Confirm, cx: &mut ViewContext<Self>) {
-        match self.codegen_status(cx) {
-            CodegenStatus::Idle => {
-                cx.emit(PromptEditorEvent::StartRequested);
-            }
-            CodegenStatus::Pending => {
-                cx.emit(PromptEditorEvent::DismissRequested);
-            }
-            CodegenStatus::Done => {
-                if self.edited_since_done {
-                    cx.emit(PromptEditorEvent::StartRequested);
-                } else {
-                    cx.emit(PromptEditorEvent::ConfirmRequested { execute: false });
-                }
-            }
-            CodegenStatus::Error(_) => {
-                cx.emit(PromptEditorEvent::StartRequested);
-            }
-        }
-    }
-
-    fn move_up(&mut self, _: &MoveUp, cx: &mut ViewContext<Self>) {
-        if let Some(ix) = self.prompt_history_ix {
-            if ix > 0 {
-                self.prompt_history_ix = Some(ix - 1);
-                let prompt = self.prompt_history[ix - 1].as_str();
-                self.editor.update(cx, |editor, cx| {
-                    editor.set_text(prompt, cx);
-                    editor.move_to_beginning(&Default::default(), cx);
-                });
-            }
-        } else if !self.prompt_history.is_empty() {
-            self.prompt_history_ix = Some(self.prompt_history.len() - 1);
-            let prompt = self.prompt_history[self.prompt_history.len() - 1].as_str();
-            self.editor.update(cx, |editor, cx| {
-                editor.set_text(prompt, cx);
-                editor.move_to_beginning(&Default::default(), cx);
-            });
-        }
-    }
-
-    fn move_down(&mut self, _: &MoveDown, cx: &mut ViewContext<Self>) {
-        if let Some(ix) = self.prompt_history_ix {
-            if ix < self.prompt_history.len() - 1 {
-                self.prompt_history_ix = Some(ix + 1);
-                let prompt = self.prompt_history[ix + 1].as_str();
-                self.editor.update(cx, |editor, cx| {
-                    editor.set_text(prompt, cx);
-                    editor.move_to_end(&Default::default(), cx)
-                });
-            } else {
-                self.prompt_history_ix = None;
-                let prompt = self.pending_prompt.as_str();
-                self.editor.update(cx, |editor, cx| {
-                    editor.set_text(prompt, cx);
-                    editor.move_to_end(&Default::default(), cx)
-                });
-            }
-        }
-    }
-
-    fn cycle_prev(&mut self, _: &CyclePreviousInlineAssist, cx: &mut ViewContext<Self>) {
-        match &self.mode {
-            PromptEditorMode::Buffer { codegen } => {
-                codegen.update(cx, |codegen, cx| codegen.cycle_prev(cx));
-            }
-            PromptEditorMode::Terminal { codegen } => todo!("terminal-prompt-cycle"),
-        }
-    }
-
-    fn cycle_next(&mut self, _: &CycleNextInlineAssist, cx: &mut ViewContext<Self>) {
-        match &self.mode {
-            PromptEditorMode::Buffer { codegen } => {
-                codegen.update(cx, |codegen, cx| codegen.cycle_prev(cx));
-            }
-            PromptEditorMode::Terminal { codegen } => todo!("terminal-prompt-cycle"),
-        }
-    }
-
-    fn render_cycle_controls(&self, cx: &ViewContext<Self>) -> AnyElement {
-        let PromptEditorMode::Buffer { codegen } = &self.mode else {
-            // todo(terminal-prompt-cycle)
-            return div().into_any_element();
-        };
-
-        let codegen = codegen.read(cx);
-        let disabled = matches!(codegen.status(cx), CodegenStatus::Idle);
-
-        let model_registry = LanguageModelRegistry::read_global(cx);
-        let default_model = model_registry.active_model();
-        let alternative_models = model_registry.inline_alternative_models();
-
-        let get_model_name = |index: usize| -> String {
-            let name = |model: &Arc<dyn LanguageModel>| model.name().0.to_string();
-
-            match index {
-                0 => default_model.as_ref().map_or_else(String::new, name),
-                index if index <= alternative_models.len() => alternative_models
-                    .get(index - 1)
-                    .map_or_else(String::new, name),
-                _ => String::new(),
-            }
-        };
-
-        let total_models = alternative_models.len() + 1;
-
-        if total_models <= 1 {
-            return div().into_any_element();
-        }
-
-        let current_index = codegen.active_alternative;
-        let prev_index = (current_index + total_models - 1) % total_models;
-        let next_index = (current_index + 1) % total_models;
-
-        let prev_model_name = get_model_name(prev_index);
-        let next_model_name = get_model_name(next_index);
-
-        h_flex()
-            .child(
-                IconButton::new("previous", IconName::ChevronLeft)
-                    .icon_color(Color::Muted)
-                    .disabled(disabled || current_index == 0)
-                    .shape(IconButtonShape::Square)
-                    .tooltip({
-                        let focus_handle = self.editor.focus_handle(cx);
-                        move |cx| {
-                            cx.new_view(|cx| {
-                                let mut tooltip = Tooltip::new("Previous Alternative").key_binding(
-                                    KeyBinding::for_action_in(
-                                        &CyclePreviousInlineAssist,
-                                        &focus_handle,
-                                        cx,
-                                    ),
-                                );
-                                if !disabled && current_index != 0 {
-                                    tooltip = tooltip.meta(prev_model_name.clone());
-                                }
-                                tooltip
-                            })
-                            .into()
-                        }
-                    })
-                    .on_click(cx.listener(|this, _, cx| {
-                        // this.codegen
-                        //     .update(cx, |codegen, cx| codegen.cycle_prev(cx))
-                        todo!("terminal-prompt-cycle")
-                    })),
-            )
-            .child(
-                Label::new(format!(
-                    "{}/{}",
-                    codegen.active_alternative + 1,
-                    codegen.alternative_count(cx)
-                ))
-                .size(LabelSize::Small)
-                .color(if disabled {
-                    Color::Disabled
-                } else {
-                    Color::Muted
-                }),
-            )
-            .child(
-                IconButton::new("next", IconName::ChevronRight)
-                    .icon_color(Color::Muted)
-                    .disabled(disabled || current_index == total_models - 1)
-                    .shape(IconButtonShape::Square)
-                    .tooltip({
-                        let focus_handle = self.editor.focus_handle(cx);
-                        move |cx| {
-                            cx.new_view(|cx| {
-                                let mut tooltip = Tooltip::new("Next Alternative").key_binding(
-                                    KeyBinding::for_action_in(
-                                        &CycleNextInlineAssist,
-                                        &focus_handle,
-                                        cx,
-                                    ),
-                                );
-                                if !disabled && current_index != total_models - 1 {
-                                    tooltip = tooltip.meta(next_model_name.clone());
-                                }
-                                tooltip
-                            })
-                            .into()
-                        }
-                    })
-                    .on_click(cx.listener(|this, _, cx| {
-                        // this.codegen
-                        //     .update(cx, |codegen, cx| codegen.cycle_next(cx))
-                        todo!("terminal-prompt-cycle")
-                    })),
-            )
-            .into_any_element()
-    }
-
-    fn render_rate_limit_notice(&self, cx: &mut ViewContext<Self>) -> impl IntoElement {
-        Popover::new().child(
-            v_flex()
-                .occlude()
-                .p_2()
-                .child(
-                    Label::new("Out of Tokens")
-                        .size(LabelSize::Small)
-                        .weight(FontWeight::BOLD),
-                )
-                .child(Label::new(
-                    "Try Zed Pro for higher limits, a wider range of models, and more.",
-                ))
-                .child(
-                    h_flex()
-                        .justify_between()
-                        .child(CheckboxWithLabel::new(
-                            "dont-show-again",
-                            Label::new("Don't show again"),
-                            if dismissed_rate_limit_notice() {
-                                ui::ToggleState::Selected
-                            } else {
-                                ui::ToggleState::Unselected
-                            },
-                            |selection, cx| {
-                                let is_dismissed = match selection {
-                                    ui::ToggleState::Unselected => false,
-                                    ui::ToggleState::Indeterminate => return,
-                                    ui::ToggleState::Selected => true,
-                                };
-
-                                set_rate_limit_notice_dismissed(is_dismissed, cx)
-                            },
-                        ))
-                        .child(
-                            h_flex()
-                                .gap_2()
-                                .child(
-                                    Button::new("dismiss", "Dismiss")
-                                        .style(ButtonStyle::Transparent)
-                                        .on_click(cx.listener(Self::toggle_rate_limit_notice)),
-                                )
-                                .child(Button::new("more-info", "More Info").on_click(
-                                    |_event, cx| {
-                                        cx.dispatch_action(Box::new(
-                                            zed_actions::OpenAccountSettings,
-                                        ))
-                                    },
-                                )),
-                        ),
-                ),
-        )
-    }
-
-    fn render_editor(&mut self, cx: &mut ViewContext<Self>) -> AnyElement {
-        let font_size = TextSize::Default.rems(cx);
-        let line_height = font_size.to_pixels(cx.rem_size()) * 1.3;
-
-        v_flex()
-            .key_context("MessageEditor")
-            .size_full()
-            .gap_2()
-            .p_2()
-            .bg(cx.theme().colors().editor_background)
-            .child({
-                let settings = ThemeSettings::get_global(cx);
-                let text_style = TextStyle {
-                    color: cx.theme().colors().editor_foreground,
-                    font_family: settings.ui_font.family.clone(),
-                    font_features: settings.ui_font.features.clone(),
-                    font_size: font_size.into(),
-                    font_weight: settings.ui_font.weight,
-                    line_height: line_height.into(),
-                    ..Default::default()
-                };
-
-                EditorElement::new(
-                    &self.editor,
-                    EditorStyle {
-                        background: cx.theme().colors().editor_background,
-                        local_player: cx.theme().players().local(),
-                        text: text_style,
-                        ..Default::default()
-                    },
-                )
-            })
-            .into_any_element()
-    }
-}
-
-const DISMISSED_RATE_LIMIT_NOTICE_KEY: &str = "dismissed-rate-limit-notice";
-
-fn dismissed_rate_limit_notice() -> bool {
-    db::kvp::KEY_VALUE_STORE
-        .read_kvp(DISMISSED_RATE_LIMIT_NOTICE_KEY)
-        .log_err()
-        .map_or(false, |s| s.is_some())
-}
-
-fn set_rate_limit_notice_dismissed(is_dismissed: bool, cx: &mut AppContext) {
-    db::write_and_log(cx, move || async move {
-        if is_dismissed {
-            db::kvp::KEY_VALUE_STORE
-                .write_kvp(DISMISSED_RATE_LIMIT_NOTICE_KEY.into(), "1".into())
-                .await
-        } else {
-            db::kvp::KEY_VALUE_STORE
-                .delete_kvp(DISMISSED_RATE_LIMIT_NOTICE_KEY.into())
-                .await
-        }
-    })
-}
-
-enum PromptEditorMode {
-    Buffer { codegen: Model<BufferCodegen> },
-    Terminal { codegen: Model<TerminalCodegen> },
 }
 
 pub struct InlineAssist {
@@ -2331,1189 +1575,6 @@ struct InlineAssistDecorations {
     prompt_editor: View<PromptEditor>,
     removed_line_block_ids: HashSet<CustomBlockId>,
     end_block_id: CustomBlockId,
-}
-
-#[derive(Copy, Clone, Debug)]
-pub enum CodegenEvent {
-    Finished,
-    Undone,
-}
-
-pub struct BufferCodegen {
-    alternatives: Vec<Model<CodegenAlternative>>,
-    active_alternative: usize,
-    seen_alternatives: HashSet<usize>,
-    subscriptions: Vec<Subscription>,
-    buffer: Model<MultiBuffer>,
-    range: Range<Anchor>,
-    initial_transaction_id: Option<TransactionId>,
-    context_store: Model<ContextStore>,
-    telemetry: Arc<Telemetry>,
-    builder: Arc<PromptBuilder>,
-    is_insertion: bool,
-}
-
-impl BufferCodegen {
-    pub fn new(
-        buffer: Model<MultiBuffer>,
-        range: Range<Anchor>,
-        initial_transaction_id: Option<TransactionId>,
-        context_store: Model<ContextStore>,
-        telemetry: Arc<Telemetry>,
-        builder: Arc<PromptBuilder>,
-        cx: &mut ModelContext<Self>,
-    ) -> Self {
-        let codegen = cx.new_model(|cx| {
-            CodegenAlternative::new(
-                buffer.clone(),
-                range.clone(),
-                false,
-                Some(context_store.clone()),
-                Some(telemetry.clone()),
-                builder.clone(),
-                cx,
-            )
-        });
-        let mut this = Self {
-            is_insertion: range.to_offset(&buffer.read(cx).snapshot(cx)).is_empty(),
-            alternatives: vec![codegen],
-            active_alternative: 0,
-            seen_alternatives: HashSet::default(),
-            subscriptions: Vec::new(),
-            buffer,
-            range,
-            initial_transaction_id,
-            context_store,
-            telemetry,
-            builder,
-        };
-        this.activate(0, cx);
-        this
-    }
-
-    fn subscribe_to_alternative(&mut self, cx: &mut ModelContext<Self>) {
-        let codegen = self.active_alternative().clone();
-        self.subscriptions.clear();
-        self.subscriptions
-            .push(cx.observe(&codegen, |_, _, cx| cx.notify()));
-        self.subscriptions
-            .push(cx.subscribe(&codegen, |_, _, event, cx| cx.emit(*event)));
-    }
-
-    fn active_alternative(&self) -> &Model<CodegenAlternative> {
-        &self.alternatives[self.active_alternative]
-    }
-
-    fn status<'a>(&self, cx: &'a AppContext) -> &'a CodegenStatus {
-        &self.active_alternative().read(cx).status
-    }
-
-    fn alternative_count(&self, cx: &AppContext) -> usize {
-        LanguageModelRegistry::read_global(cx)
-            .inline_alternative_models()
-            .len()
-            + 1
-    }
-
-    pub fn cycle_prev(&mut self, cx: &mut ModelContext<Self>) {
-        let next_active_ix = if self.active_alternative == 0 {
-            self.alternatives.len() - 1
-        } else {
-            self.active_alternative - 1
-        };
-        self.activate(next_active_ix, cx);
-    }
-
-    pub fn cycle_next(&mut self, cx: &mut ModelContext<Self>) {
-        let next_active_ix = (self.active_alternative + 1) % self.alternatives.len();
-        self.activate(next_active_ix, cx);
-    }
-
-    fn activate(&mut self, index: usize, cx: &mut ModelContext<Self>) {
-        self.active_alternative()
-            .update(cx, |codegen, cx| codegen.set_active(false, cx));
-        self.seen_alternatives.insert(index);
-        self.active_alternative = index;
-        self.active_alternative()
-            .update(cx, |codegen, cx| codegen.set_active(true, cx));
-        self.subscribe_to_alternative(cx);
-        cx.notify();
-    }
-
-    pub fn start(&mut self, user_prompt: String, cx: &mut ModelContext<Self>) -> Result<()> {
-        let alternative_models = LanguageModelRegistry::read_global(cx)
-            .inline_alternative_models()
-            .to_vec();
-
-        self.active_alternative()
-            .update(cx, |alternative, cx| alternative.undo(cx));
-        self.activate(0, cx);
-        self.alternatives.truncate(1);
-
-        for _ in 0..alternative_models.len() {
-            self.alternatives.push(cx.new_model(|cx| {
-                CodegenAlternative::new(
-                    self.buffer.clone(),
-                    self.range.clone(),
-                    false,
-                    Some(self.context_store.clone()),
-                    Some(self.telemetry.clone()),
-                    self.builder.clone(),
-                    cx,
-                )
-            }));
-        }
-
-        let primary_model = LanguageModelRegistry::read_global(cx)
-            .active_model()
-            .context("no active model")?;
-
-        for (model, alternative) in iter::once(primary_model)
-            .chain(alternative_models)
-            .zip(&self.alternatives)
-        {
-            alternative.update(cx, |alternative, cx| {
-                alternative.start(user_prompt.clone(), model.clone(), cx)
-            })?;
-        }
-
-        Ok(())
-    }
-
-    pub fn stop(&mut self, cx: &mut ModelContext<Self>) {
-        for codegen in &self.alternatives {
-            codegen.update(cx, |codegen, cx| codegen.stop(cx));
-        }
-    }
-
-    pub fn undo(&mut self, cx: &mut ModelContext<Self>) {
-        self.active_alternative()
-            .update(cx, |codegen, cx| codegen.undo(cx));
-
-        self.buffer.update(cx, |buffer, cx| {
-            if let Some(transaction_id) = self.initial_transaction_id.take() {
-                buffer.undo_transaction(transaction_id, cx);
-                buffer.refresh_preview(cx);
-            }
-        });
-    }
-
-    pub fn buffer(&self, cx: &AppContext) -> Model<MultiBuffer> {
-        self.active_alternative().read(cx).buffer.clone()
-    }
-
-    pub fn old_buffer(&self, cx: &AppContext) -> Model<Buffer> {
-        self.active_alternative().read(cx).old_buffer.clone()
-    }
-
-    pub fn snapshot(&self, cx: &AppContext) -> MultiBufferSnapshot {
-        self.active_alternative().read(cx).snapshot.clone()
-    }
-
-    pub fn edit_position(&self, cx: &AppContext) -> Option<Anchor> {
-        self.active_alternative().read(cx).edit_position
-    }
-
-    fn diff<'a>(&self, cx: &'a AppContext) -> &'a Diff {
-        &self.active_alternative().read(cx).diff
-    }
-
-    pub fn last_equal_ranges<'a>(&self, cx: &'a AppContext) -> &'a [Range<Anchor>] {
-        self.active_alternative().read(cx).last_equal_ranges()
-    }
-}
-
-impl EventEmitter<CodegenEvent> for BufferCodegen {}
-
-pub struct CodegenAlternative {
-    buffer: Model<MultiBuffer>,
-    old_buffer: Model<Buffer>,
-    snapshot: MultiBufferSnapshot,
-    edit_position: Option<Anchor>,
-    range: Range<Anchor>,
-    last_equal_ranges: Vec<Range<Anchor>>,
-    transformation_transaction_id: Option<TransactionId>,
-    status: CodegenStatus,
-    generation: Task<()>,
-    diff: Diff,
-    context_store: Option<Model<ContextStore>>,
-    telemetry: Option<Arc<Telemetry>>,
-    _subscription: gpui::Subscription,
-    builder: Arc<PromptBuilder>,
-    active: bool,
-    edits: Vec<(Range<Anchor>, String)>,
-    line_operations: Vec<LineOperation>,
-    request: Option<LanguageModelRequest>,
-    elapsed_time: Option<f64>,
-    completion: Option<String>,
-    message_id: Option<String>,
-}
-
-const CLEAR_INPUT: &str = "\x15";
-const CARRIAGE_RETURN: &str = "\x0d";
-
-struct TerminalTransaction {
-    terminal: Model<Terminal>,
-}
-
-impl TerminalTransaction {
-    pub fn start(terminal: Model<Terminal>) -> Self {
-        Self { terminal }
-    }
-
-    pub fn push(&mut self, hunk: String, cx: &mut AppContext) {
-        // Ensure that the assistant cannot accidentally execute commands that are streamed into the terminal
-        let input = Self::sanitize_input(hunk);
-        self.terminal
-            .update(cx, |terminal, _| terminal.input(input));
-    }
-
-    pub fn undo(&self, cx: &mut AppContext) {
-        self.terminal
-            .update(cx, |terminal, _| terminal.input(CLEAR_INPUT.to_string()));
-    }
-
-    pub fn complete(&self, cx: &mut AppContext) {
-        self.terminal.update(cx, |terminal, _| {
-            terminal.input(CARRIAGE_RETURN.to_string())
-        });
-    }
-
-    fn sanitize_input(input: String) -> String {
-        input.replace(['\r', '\n'], "")
-    }
-}
-
-impl EventEmitter<CodegenEvent> for TerminalCodegen {}
-
-pub struct TerminalCodegen {
-    status: CodegenStatus,
-    telemetry: Option<Arc<Telemetry>>,
-    terminal: Model<Terminal>,
-    generation: Task<()>,
-    message_id: Option<String>,
-    transaction: Option<TerminalTransaction>,
-}
-
-impl TerminalCodegen {
-    pub fn new(terminal: Model<Terminal>, telemetry: Option<Arc<Telemetry>>) -> Self {
-        Self {
-            terminal,
-            telemetry,
-            status: CodegenStatus::Idle,
-            generation: Task::ready(()),
-            message_id: None,
-            transaction: None,
-        }
-    }
-
-    pub fn start(&mut self, prompt: LanguageModelRequest, cx: &mut ModelContext<Self>) {
-        let Some(model) = LanguageModelRegistry::read_global(cx).active_model() else {
-            return;
-        };
-
-        let model_api_key = model.api_key(cx);
-        let http_client = cx.http_client();
-        let telemetry = self.telemetry.clone();
-        self.status = CodegenStatus::Pending;
-        self.transaction = Some(TerminalTransaction::start(self.terminal.clone()));
-        self.generation = cx.spawn(|this, mut cx| async move {
-            let model_telemetry_id = model.telemetry_id();
-            let model_provider_id = model.provider_id();
-            let response = model.stream_completion_text(prompt, &cx).await;
-            let generate = async {
-                let message_id = response
-                    .as_ref()
-                    .ok()
-                    .and_then(|response| response.message_id.clone());
-
-                let (mut hunks_tx, mut hunks_rx) = mpsc::channel(1);
-
-                let task = cx.background_executor().spawn({
-                    let message_id = message_id.clone();
-                    let executor = cx.background_executor().clone();
-                    async move {
-                        let mut response_latency = None;
-                        let request_start = Instant::now();
-                        let task = async {
-                            let mut chunks = response?.stream;
-                            while let Some(chunk) = chunks.next().await {
-                                if response_latency.is_none() {
-                                    response_latency = Some(request_start.elapsed());
-                                }
-                                let chunk = chunk?;
-                                hunks_tx.send(chunk).await?;
-                            }
-
-                            anyhow::Ok(())
-                        };
-
-                        let result = task.await;
-
-                        let error_message = result.as_ref().err().map(|error| error.to_string());
-                        report_assistant_event(
-                            AssistantEvent {
-                                conversation_id: None,
-                                kind: AssistantKind::InlineTerminal,
-                                message_id,
-                                phase: AssistantPhase::Response,
-                                model: model_telemetry_id,
-                                model_provider: model_provider_id.to_string(),
-                                response_latency,
-                                error_message,
-                                language_name: None,
-                            },
-                            telemetry,
-                            http_client,
-                            model_api_key,
-                            &executor,
-                        );
-
-                        result?;
-                        anyhow::Ok(())
-                    }
-                });
-
-                this.update(&mut cx, |this, _| {
-                    this.message_id = message_id;
-                })?;
-
-                while let Some(hunk) = hunks_rx.next().await {
-                    this.update(&mut cx, |this, cx| {
-                        if let Some(transaction) = &mut this.transaction {
-                            transaction.push(hunk, cx);
-                            cx.notify();
-                        }
-                    })?;
-                }
-
-                task.await?;
-                anyhow::Ok(())
-            };
-
-            let result = generate.await;
-
-            this.update(&mut cx, |this, cx| {
-                if let Err(error) = result {
-                    this.status = CodegenStatus::Error(error);
-                } else {
-                    this.status = CodegenStatus::Done;
-                }
-                cx.emit(CodegenEvent::Finished);
-                cx.notify();
-            })
-            .ok();
-        });
-        cx.notify();
-    }
-
-    pub fn stop(&mut self, cx: &mut ModelContext<Self>) {
-        self.status = CodegenStatus::Done;
-        self.generation = Task::ready(());
-        cx.emit(CodegenEvent::Finished);
-        cx.notify();
-    }
-
-    pub fn complete(&mut self, cx: &mut ModelContext<Self>) {
-        if let Some(transaction) = self.transaction.take() {
-            transaction.complete(cx);
-        }
-    }
-
-    pub fn undo(&mut self, cx: &mut ModelContext<Self>) {
-        if let Some(transaction) = self.transaction.take() {
-            transaction.undo(cx);
-        }
-    }
-}
-
-#[derive(Default)]
-struct Diff {
-    deleted_row_ranges: Vec<(Anchor, RangeInclusive<u32>)>,
-    inserted_row_ranges: Vec<Range<Anchor>>,
-}
-
-impl Diff {
-    fn is_empty(&self) -> bool {
-        self.deleted_row_ranges.is_empty() && self.inserted_row_ranges.is_empty()
-    }
-}
-
-impl EventEmitter<CodegenEvent> for CodegenAlternative {}
-
-impl CodegenAlternative {
-    pub fn new(
-        buffer: Model<MultiBuffer>,
-        range: Range<Anchor>,
-        active: bool,
-        context_store: Option<Model<ContextStore>>,
-        telemetry: Option<Arc<Telemetry>>,
-        builder: Arc<PromptBuilder>,
-        cx: &mut ModelContext<Self>,
-    ) -> Self {
-        let snapshot = buffer.read(cx).snapshot(cx);
-
-        let (old_buffer, _, _) = buffer
-            .read(cx)
-            .range_to_buffer_ranges(range.clone(), cx)
-            .pop()
-            .unwrap();
-        let old_buffer = cx.new_model(|cx| {
-            let old_buffer = old_buffer.read(cx);
-            let text = old_buffer.as_rope().clone();
-            let line_ending = old_buffer.line_ending();
-            let language = old_buffer.language().cloned();
-            let language_registry = old_buffer.language_registry();
-
-            let mut buffer = Buffer::local_normalized(text, line_ending, cx);
-            buffer.set_language(language, cx);
-            if let Some(language_registry) = language_registry {
-                buffer.set_language_registry(language_registry)
-            }
-            buffer
-        });
-
-        Self {
-            buffer: buffer.clone(),
-            old_buffer,
-            edit_position: None,
-            message_id: None,
-            snapshot,
-            last_equal_ranges: Default::default(),
-            transformation_transaction_id: None,
-            status: CodegenStatus::Idle,
-            generation: Task::ready(()),
-            diff: Diff::default(),
-            context_store,
-            telemetry,
-            _subscription: cx.subscribe(&buffer, Self::handle_buffer_event),
-            builder,
-            active,
-            edits: Vec::new(),
-            line_operations: Vec::new(),
-            range,
-            request: None,
-            elapsed_time: None,
-            completion: None,
-        }
-    }
-
-    fn set_active(&mut self, active: bool, cx: &mut ModelContext<Self>) {
-        if active != self.active {
-            self.active = active;
-
-            if self.active {
-                let edits = self.edits.clone();
-                self.apply_edits(edits, cx);
-                if matches!(self.status, CodegenStatus::Pending) {
-                    let line_operations = self.line_operations.clone();
-                    self.reapply_line_based_diff(line_operations, cx);
-                } else {
-                    self.reapply_batch_diff(cx).detach();
-                }
-            } else if let Some(transaction_id) = self.transformation_transaction_id.take() {
-                self.buffer.update(cx, |buffer, cx| {
-                    buffer.undo_transaction(transaction_id, cx);
-                    buffer.forget_transaction(transaction_id, cx);
-                });
-            }
-        }
-    }
-
-    fn handle_buffer_event(
-        &mut self,
-        _buffer: Model<MultiBuffer>,
-        event: &multi_buffer::Event,
-        cx: &mut ModelContext<Self>,
-    ) {
-        if let multi_buffer::Event::TransactionUndone { transaction_id } = event {
-            if self.transformation_transaction_id == Some(*transaction_id) {
-                self.transformation_transaction_id = None;
-                self.generation = Task::ready(());
-                cx.emit(CodegenEvent::Undone);
-            }
-        }
-    }
-
-    pub fn last_equal_ranges(&self) -> &[Range<Anchor>] {
-        &self.last_equal_ranges
-    }
-
-    pub fn start(
-        &mut self,
-        user_prompt: String,
-        model: Arc<dyn LanguageModel>,
-        cx: &mut ModelContext<Self>,
-    ) -> Result<()> {
-        if let Some(transformation_transaction_id) = self.transformation_transaction_id.take() {
-            self.buffer.update(cx, |buffer, cx| {
-                buffer.undo_transaction(transformation_transaction_id, cx);
-            });
-        }
-
-        self.edit_position = Some(self.range.start.bias_right(&self.snapshot));
-
-        let api_key = model.api_key(cx);
-        let telemetry_id = model.telemetry_id();
-        let provider_id = model.provider_id();
-        let stream: LocalBoxFuture<Result<LanguageModelTextStream>> =
-            if user_prompt.trim().to_lowercase() == "delete" {
-                async { Ok(LanguageModelTextStream::default()) }.boxed_local()
-            } else {
-                let request = self.build_request(user_prompt, cx)?;
-                self.request = Some(request.clone());
-
-                cx.spawn(|_, cx| async move { model.stream_completion_text(request, &cx).await })
-                    .boxed_local()
-            };
-        self.handle_stream(telemetry_id, provider_id.to_string(), api_key, stream, cx);
-        Ok(())
-    }
-
-    fn build_request(
-        &self,
-        user_prompt: String,
-        cx: &mut AppContext,
-    ) -> Result<LanguageModelRequest> {
-        let buffer = self.buffer.read(cx).snapshot(cx);
-        let language = buffer.language_at(self.range.start);
-        let language_name = if let Some(language) = language.as_ref() {
-            if Arc::ptr_eq(language, &language::PLAIN_TEXT) {
-                None
-            } else {
-                Some(language.name())
-            }
-        } else {
-            None
-        };
-
-        let language_name = language_name.as_ref();
-        let start = buffer.point_to_buffer_offset(self.range.start);
-        let end = buffer.point_to_buffer_offset(self.range.end);
-        let (buffer, range) = if let Some((start, end)) = start.zip(end) {
-            let (start_buffer, start_buffer_offset) = start;
-            let (end_buffer, end_buffer_offset) = end;
-            if start_buffer.remote_id() == end_buffer.remote_id() {
-                (start_buffer.clone(), start_buffer_offset..end_buffer_offset)
-            } else {
-                return Err(anyhow::anyhow!("invalid transformation range"));
-            }
-        } else {
-            return Err(anyhow::anyhow!("invalid transformation range"));
-        };
-
-        let prompt = self
-            .builder
-            .generate_inline_transformation_prompt(user_prompt, language_name, buffer, range)
-            .map_err(|e| anyhow::anyhow!("Failed to generate content prompt: {}", e))?;
-
-        let mut request_message = LanguageModelRequestMessage {
-            role: Role::User,
-            content: Vec::new(),
-            cache: false,
-        };
-
-        if let Some(context_store) = &self.context_store {
-            let context = context_store.update(cx, |this, _cx| this.context().clone());
-            attach_context_to_message(&mut request_message, context);
-        }
-
-        request_message.content.push(prompt.into());
-
-        Ok(LanguageModelRequest {
-            tools: Vec::new(),
-            stop: Vec::new(),
-            temperature: None,
-            messages: vec![request_message],
-        })
-    }
-
-    pub fn handle_stream(
-        &mut self,
-        model_telemetry_id: String,
-        model_provider_id: String,
-        model_api_key: Option<String>,
-        stream: impl 'static + Future<Output = Result<LanguageModelTextStream>>,
-        cx: &mut ModelContext<Self>,
-    ) {
-        let start_time = Instant::now();
-        let snapshot = self.snapshot.clone();
-        let selected_text = snapshot
-            .text_for_range(self.range.start..self.range.end)
-            .collect::<Rope>();
-
-        let selection_start = self.range.start.to_point(&snapshot);
-
-        // Start with the indentation of the first line in the selection
-        let mut suggested_line_indent = snapshot
-            .suggested_indents(selection_start.row..=selection_start.row, cx)
-            .into_values()
-            .next()
-            .unwrap_or_else(|| snapshot.indent_size_for_line(MultiBufferRow(selection_start.row)));
-
-        // If the first line in the selection does not have indentation, check the following lines
-        if suggested_line_indent.len == 0 && suggested_line_indent.kind == IndentKind::Space {
-            for row in selection_start.row..=self.range.end.to_point(&snapshot).row {
-                let line_indent = snapshot.indent_size_for_line(MultiBufferRow(row));
-                // Prefer tabs if a line in the selection uses tabs as indentation
-                if line_indent.kind == IndentKind::Tab {
-                    suggested_line_indent.kind = IndentKind::Tab;
-                    break;
-                }
-            }
-        }
-
-        let http_client = cx.http_client().clone();
-        let telemetry = self.telemetry.clone();
-        let language_name = {
-            let multibuffer = self.buffer.read(cx);
-            let ranges = multibuffer.range_to_buffer_ranges(self.range.clone(), cx);
-            ranges
-                .first()
-                .and_then(|(buffer, _, _)| buffer.read(cx).language())
-                .map(|language| language.name())
-        };
-
-        self.diff = Diff::default();
-        self.status = CodegenStatus::Pending;
-        let mut edit_start = self.range.start.to_offset(&snapshot);
-        let completion = Arc::new(Mutex::new(String::new()));
-        let completion_clone = completion.clone();
-
-        self.generation = cx.spawn(|codegen, mut cx| {
-            async move {
-                let stream = stream.await;
-                let message_id = stream
-                    .as_ref()
-                    .ok()
-                    .and_then(|stream| stream.message_id.clone());
-                let generate = async {
-                    let (mut diff_tx, mut diff_rx) = mpsc::channel(1);
-                    let executor = cx.background_executor().clone();
-                    let message_id = message_id.clone();
-                    let line_based_stream_diff: Task<anyhow::Result<()>> =
-                        cx.background_executor().spawn(async move {
-                            let mut response_latency = None;
-                            let request_start = Instant::now();
-                            let diff = async {
-                                let chunks = StripInvalidSpans::new(stream?.stream);
-                                futures::pin_mut!(chunks);
-                                let mut diff = StreamingDiff::new(selected_text.to_string());
-                                let mut line_diff = LineDiff::default();
-
-                                let mut new_text = String::new();
-                                let mut base_indent = None;
-                                let mut line_indent = None;
-                                let mut first_line = true;
-
-                                while let Some(chunk) = chunks.next().await {
-                                    if response_latency.is_none() {
-                                        response_latency = Some(request_start.elapsed());
-                                    }
-                                    let chunk = chunk?;
-                                    completion_clone.lock().push_str(&chunk);
-
-                                    let mut lines = chunk.split('\n').peekable();
-                                    while let Some(line) = lines.next() {
-                                        new_text.push_str(line);
-                                        if line_indent.is_none() {
-                                            if let Some(non_whitespace_ch_ix) =
-                                                new_text.find(|ch: char| !ch.is_whitespace())
-                                            {
-                                                line_indent = Some(non_whitespace_ch_ix);
-                                                base_indent = base_indent.or(line_indent);
-
-                                                let line_indent = line_indent.unwrap();
-                                                let base_indent = base_indent.unwrap();
-                                                let indent_delta =
-                                                    line_indent as i32 - base_indent as i32;
-                                                let mut corrected_indent_len = cmp::max(
-                                                    0,
-                                                    suggested_line_indent.len as i32 + indent_delta,
-                                                )
-                                                    as usize;
-                                                if first_line {
-                                                    corrected_indent_len = corrected_indent_len
-                                                        .saturating_sub(
-                                                            selection_start.column as usize,
-                                                        );
-                                                }
-
-                                                let indent_char = suggested_line_indent.char();
-                                                let mut indent_buffer = [0; 4];
-                                                let indent_str =
-                                                    indent_char.encode_utf8(&mut indent_buffer);
-                                                new_text.replace_range(
-                                                    ..line_indent,
-                                                    &indent_str.repeat(corrected_indent_len),
-                                                );
-                                            }
-                                        }
-
-                                        if line_indent.is_some() {
-                                            let char_ops = diff.push_new(&new_text);
-                                            line_diff
-                                                .push_char_operations(&char_ops, &selected_text);
-                                            diff_tx
-                                                .send((char_ops, line_diff.line_operations()))
-                                                .await?;
-                                            new_text.clear();
-                                        }
-
-                                        if lines.peek().is_some() {
-                                            let char_ops = diff.push_new("\n");
-                                            line_diff
-                                                .push_char_operations(&char_ops, &selected_text);
-                                            diff_tx
-                                                .send((char_ops, line_diff.line_operations()))
-                                                .await?;
-                                            if line_indent.is_none() {
-                                                // Don't write out the leading indentation in empty lines on the next line
-                                                // This is the case where the above if statement didn't clear the buffer
-                                                new_text.clear();
-                                            }
-                                            line_indent = None;
-                                            first_line = false;
-                                        }
-                                    }
-                                }
-
-                                let mut char_ops = diff.push_new(&new_text);
-                                char_ops.extend(diff.finish());
-                                line_diff.push_char_operations(&char_ops, &selected_text);
-                                line_diff.finish(&selected_text);
-                                diff_tx
-                                    .send((char_ops, line_diff.line_operations()))
-                                    .await?;
-
-                                anyhow::Ok(())
-                            };
-
-                            let result = diff.await;
-
-                            let error_message =
-                                result.as_ref().err().map(|error| error.to_string());
-                            report_assistant_event(
-                                AssistantEvent {
-                                    conversation_id: None,
-                                    message_id,
-                                    kind: AssistantKind::Inline,
-                                    phase: AssistantPhase::Response,
-                                    model: model_telemetry_id,
-                                    model_provider: model_provider_id.to_string(),
-                                    response_latency,
-                                    error_message,
-                                    language_name: language_name.map(|name| name.to_proto()),
-                                },
-                                telemetry,
-                                http_client,
-                                model_api_key,
-                                &executor,
-                            );
-
-                            result?;
-                            Ok(())
-                        });
-
-                    while let Some((char_ops, line_ops)) = diff_rx.next().await {
-                        codegen.update(&mut cx, |codegen, cx| {
-                            codegen.last_equal_ranges.clear();
-
-                            let edits = char_ops
-                                .into_iter()
-                                .filter_map(|operation| match operation {
-                                    CharOperation::Insert { text } => {
-                                        let edit_start = snapshot.anchor_after(edit_start);
-                                        Some((edit_start..edit_start, text))
-                                    }
-                                    CharOperation::Delete { bytes } => {
-                                        let edit_end = edit_start + bytes;
-                                        let edit_range = snapshot.anchor_after(edit_start)
-                                            ..snapshot.anchor_before(edit_end);
-                                        edit_start = edit_end;
-                                        Some((edit_range, String::new()))
-                                    }
-                                    CharOperation::Keep { bytes } => {
-                                        let edit_end = edit_start + bytes;
-                                        let edit_range = snapshot.anchor_after(edit_start)
-                                            ..snapshot.anchor_before(edit_end);
-                                        edit_start = edit_end;
-                                        codegen.last_equal_ranges.push(edit_range);
-                                        None
-                                    }
-                                })
-                                .collect::<Vec<_>>();
-
-                            if codegen.active {
-                                codegen.apply_edits(edits.iter().cloned(), cx);
-                                codegen.reapply_line_based_diff(line_ops.iter().cloned(), cx);
-                            }
-                            codegen.edits.extend(edits);
-                            codegen.line_operations = line_ops;
-                            codegen.edit_position = Some(snapshot.anchor_after(edit_start));
-
-                            cx.notify();
-                        })?;
-                    }
-
-                    // Streaming stopped and we have the new text in the buffer, and a line-based diff applied for the whole new buffer.
-                    // That diff is not what a regular diff is and might look unexpected, ergo apply a regular diff.
-                    // It's fine to apply even if the rest of the line diffing fails, as no more hunks are coming through `diff_rx`.
-                    let batch_diff_task =
-                        codegen.update(&mut cx, |codegen, cx| codegen.reapply_batch_diff(cx))?;
-                    let (line_based_stream_diff, ()) =
-                        join!(line_based_stream_diff, batch_diff_task);
-                    line_based_stream_diff?;
-
-                    anyhow::Ok(())
-                };
-
-                let result = generate.await;
-                let elapsed_time = start_time.elapsed().as_secs_f64();
-
-                codegen
-                    .update(&mut cx, |this, cx| {
-                        this.message_id = message_id;
-                        this.last_equal_ranges.clear();
-                        if let Err(error) = result {
-                            this.status = CodegenStatus::Error(error);
-                        } else {
-                            this.status = CodegenStatus::Done;
-                        }
-                        this.elapsed_time = Some(elapsed_time);
-                        this.completion = Some(completion.lock().clone());
-                        cx.emit(CodegenEvent::Finished);
-                        cx.notify();
-                    })
-                    .ok();
-            }
-        });
-        cx.notify();
-    }
-
-    pub fn stop(&mut self, cx: &mut ModelContext<Self>) {
-        self.last_equal_ranges.clear();
-        if self.diff.is_empty() {
-            self.status = CodegenStatus::Idle;
-        } else {
-            self.status = CodegenStatus::Done;
-        }
-        self.generation = Task::ready(());
-        cx.emit(CodegenEvent::Finished);
-        cx.notify();
-    }
-
-    pub fn undo(&mut self, cx: &mut ModelContext<Self>) {
-        self.buffer.update(cx, |buffer, cx| {
-            if let Some(transaction_id) = self.transformation_transaction_id.take() {
-                buffer.undo_transaction(transaction_id, cx);
-                buffer.refresh_preview(cx);
-            }
-        });
-    }
-
-    fn apply_edits(
-        &mut self,
-        edits: impl IntoIterator<Item = (Range<Anchor>, String)>,
-        cx: &mut ModelContext<CodegenAlternative>,
-    ) {
-        let transaction = self.buffer.update(cx, |buffer, cx| {
-            // Avoid grouping assistant edits with user edits.
-            buffer.finalize_last_transaction(cx);
-            buffer.start_transaction(cx);
-            buffer.edit(edits, None, cx);
-            buffer.end_transaction(cx)
-        });
-
-        if let Some(transaction) = transaction {
-            if let Some(first_transaction) = self.transformation_transaction_id {
-                // Group all assistant edits into the first transaction.
-                self.buffer.update(cx, |buffer, cx| {
-                    buffer.merge_transactions(transaction, first_transaction, cx)
-                });
-            } else {
-                self.transformation_transaction_id = Some(transaction);
-                self.buffer
-                    .update(cx, |buffer, cx| buffer.finalize_last_transaction(cx));
-            }
-        }
-    }
-
-    fn reapply_line_based_diff(
-        &mut self,
-        line_operations: impl IntoIterator<Item = LineOperation>,
-        cx: &mut ModelContext<Self>,
-    ) {
-        let old_snapshot = self.snapshot.clone();
-        let old_range = self.range.to_point(&old_snapshot);
-        let new_snapshot = self.buffer.read(cx).snapshot(cx);
-        let new_range = self.range.to_point(&new_snapshot);
-
-        let mut old_row = old_range.start.row;
-        let mut new_row = new_range.start.row;
-
-        self.diff.deleted_row_ranges.clear();
-        self.diff.inserted_row_ranges.clear();
-        for operation in line_operations {
-            match operation {
-                LineOperation::Keep { lines } => {
-                    old_row += lines;
-                    new_row += lines;
-                }
-                LineOperation::Delete { lines } => {
-                    let old_end_row = old_row + lines - 1;
-                    let new_row = new_snapshot.anchor_before(Point::new(new_row, 0));
-
-                    if let Some((_, last_deleted_row_range)) =
-                        self.diff.deleted_row_ranges.last_mut()
-                    {
-                        if *last_deleted_row_range.end() + 1 == old_row {
-                            *last_deleted_row_range = *last_deleted_row_range.start()..=old_end_row;
-                        } else {
-                            self.diff
-                                .deleted_row_ranges
-                                .push((new_row, old_row..=old_end_row));
-                        }
-                    } else {
-                        self.diff
-                            .deleted_row_ranges
-                            .push((new_row, old_row..=old_end_row));
-                    }
-
-                    old_row += lines;
-                }
-                LineOperation::Insert { lines } => {
-                    let new_end_row = new_row + lines - 1;
-                    let start = new_snapshot.anchor_before(Point::new(new_row, 0));
-                    let end = new_snapshot.anchor_before(Point::new(
-                        new_end_row,
-                        new_snapshot.line_len(MultiBufferRow(new_end_row)),
-                    ));
-                    self.diff.inserted_row_ranges.push(start..end);
-                    new_row += lines;
-                }
-            }
-
-            cx.notify();
-        }
-    }
-
-    fn reapply_batch_diff(&mut self, cx: &mut ModelContext<Self>) -> Task<()> {
-        let old_snapshot = self.snapshot.clone();
-        let old_range = self.range.to_point(&old_snapshot);
-        let new_snapshot = self.buffer.read(cx).snapshot(cx);
-        let new_range = self.range.to_point(&new_snapshot);
-
-        cx.spawn(|codegen, mut cx| async move {
-            let (deleted_row_ranges, inserted_row_ranges) = cx
-                .background_executor()
-                .spawn(async move {
-                    let old_text = old_snapshot
-                        .text_for_range(
-                            Point::new(old_range.start.row, 0)
-                                ..Point::new(
-                                    old_range.end.row,
-                                    old_snapshot.line_len(MultiBufferRow(old_range.end.row)),
-                                ),
-                        )
-                        .collect::<String>();
-                    let new_text = new_snapshot
-                        .text_for_range(
-                            Point::new(new_range.start.row, 0)
-                                ..Point::new(
-                                    new_range.end.row,
-                                    new_snapshot.line_len(MultiBufferRow(new_range.end.row)),
-                                ),
-                        )
-                        .collect::<String>();
-
-                    let mut old_row = old_range.start.row;
-                    let mut new_row = new_range.start.row;
-                    let batch_diff =
-                        similar::TextDiff::from_lines(old_text.as_str(), new_text.as_str());
-
-                    let mut deleted_row_ranges: Vec<(Anchor, RangeInclusive<u32>)> = Vec::new();
-                    let mut inserted_row_ranges = Vec::new();
-                    for change in batch_diff.iter_all_changes() {
-                        let line_count = change.value().lines().count() as u32;
-                        match change.tag() {
-                            similar::ChangeTag::Equal => {
-                                old_row += line_count;
-                                new_row += line_count;
-                            }
-                            similar::ChangeTag::Delete => {
-                                let old_end_row = old_row + line_count - 1;
-                                let new_row = new_snapshot.anchor_before(Point::new(new_row, 0));
-
-                                if let Some((_, last_deleted_row_range)) =
-                                    deleted_row_ranges.last_mut()
-                                {
-                                    if *last_deleted_row_range.end() + 1 == old_row {
-                                        *last_deleted_row_range =
-                                            *last_deleted_row_range.start()..=old_end_row;
-                                    } else {
-                                        deleted_row_ranges.push((new_row, old_row..=old_end_row));
-                                    }
-                                } else {
-                                    deleted_row_ranges.push((new_row, old_row..=old_end_row));
-                                }
-
-                                old_row += line_count;
-                            }
-                            similar::ChangeTag::Insert => {
-                                let new_end_row = new_row + line_count - 1;
-                                let start = new_snapshot.anchor_before(Point::new(new_row, 0));
-                                let end = new_snapshot.anchor_before(Point::new(
-                                    new_end_row,
-                                    new_snapshot.line_len(MultiBufferRow(new_end_row)),
-                                ));
-                                inserted_row_ranges.push(start..end);
-                                new_row += line_count;
-                            }
-                        }
-                    }
-
-                    (deleted_row_ranges, inserted_row_ranges)
-                })
-                .await;
-
-            codegen
-                .update(&mut cx, |codegen, cx| {
-                    codegen.diff.deleted_row_ranges = deleted_row_ranges;
-                    codegen.diff.inserted_row_ranges = inserted_row_ranges;
-                    cx.notify();
-                })
-                .ok();
-        })
-    }
-}
-
-struct StripInvalidSpans<T> {
-    stream: T,
-    stream_done: bool,
-    buffer: String,
-    first_line: bool,
-    line_end: bool,
-    starts_with_code_block: bool,
-}
-
-impl<T> StripInvalidSpans<T>
-where
-    T: Stream<Item = Result<String>>,
-{
-    fn new(stream: T) -> Self {
-        Self {
-            stream,
-            stream_done: false,
-            buffer: String::new(),
-            first_line: true,
-            line_end: false,
-            starts_with_code_block: false,
-        }
-    }
-}
-
-impl<T> Stream for StripInvalidSpans<T>
-where
-    T: Stream<Item = Result<String>>,
-{
-    type Item = Result<String>;
-
-    fn poll_next(self: Pin<&mut Self>, cx: &mut task::Context) -> Poll<Option<Self::Item>> {
-        const CODE_BLOCK_DELIMITER: &str = "```";
-        const CURSOR_SPAN: &str = "<|CURSOR|>";
-
-        let this = unsafe { self.get_unchecked_mut() };
-        loop {
-            if !this.stream_done {
-                let mut stream = unsafe { Pin::new_unchecked(&mut this.stream) };
-                match stream.as_mut().poll_next(cx) {
-                    Poll::Ready(Some(Ok(chunk))) => {
-                        this.buffer.push_str(&chunk);
-                    }
-                    Poll::Ready(Some(Err(error))) => return Poll::Ready(Some(Err(error))),
-                    Poll::Ready(None) => {
-                        this.stream_done = true;
-                    }
-                    Poll::Pending => return Poll::Pending,
-                }
-            }
-
-            let mut chunk = String::new();
-            let mut consumed = 0;
-            if !this.buffer.is_empty() {
-                let mut lines = this.buffer.split('\n').enumerate().peekable();
-                while let Some((line_ix, line)) = lines.next() {
-                    if line_ix > 0 {
-                        this.first_line = false;
-                    }
-
-                    if this.first_line {
-                        let trimmed_line = line.trim();
-                        if lines.peek().is_some() {
-                            if trimmed_line.starts_with(CODE_BLOCK_DELIMITER) {
-                                consumed += line.len() + 1;
-                                this.starts_with_code_block = true;
-                                continue;
-                            }
-                        } else if trimmed_line.is_empty()
-                            || prefixes(CODE_BLOCK_DELIMITER)
-                                .any(|prefix| trimmed_line.starts_with(prefix))
-                        {
-                            break;
-                        }
-                    }
-
-                    let line_without_cursor = line.replace(CURSOR_SPAN, "");
-                    if lines.peek().is_some() {
-                        if this.line_end {
-                            chunk.push('\n');
-                        }
-
-                        chunk.push_str(&line_without_cursor);
-                        this.line_end = true;
-                        consumed += line.len() + 1;
-                    } else if this.stream_done {
-                        if !this.starts_with_code_block
-                            || !line_without_cursor.trim().ends_with(CODE_BLOCK_DELIMITER)
-                        {
-                            if this.line_end {
-                                chunk.push('\n');
-                            }
-
-                            chunk.push_str(&line);
-                        }
-
-                        consumed += line.len();
-                    } else {
-                        let trimmed_line = line.trim();
-                        if trimmed_line.is_empty()
-                            || prefixes(CURSOR_SPAN).any(|prefix| trimmed_line.ends_with(prefix))
-                            || prefixes(CODE_BLOCK_DELIMITER)
-                                .any(|prefix| trimmed_line.ends_with(prefix))
-                        {
-                            break;
-                        } else {
-                            if this.line_end {
-                                chunk.push('\n');
-                                this.line_end = false;
-                            }
-
-                            chunk.push_str(&line_without_cursor);
-                            consumed += line.len();
-                        }
-                    }
-                }
-            }
-
-            this.buffer = this.buffer.split_off(consumed);
-            if !chunk.is_empty() {
-                return Poll::Ready(Some(Ok(chunk)));
-            } else if this.stream_done {
-                return Poll::Ready(None);
-            }
-        }
-    }
 }
 
 struct AssistantCodeActionProvider {
@@ -3643,10 +1704,6 @@ impl CodeActionProvider for AssistantCodeActionProvider {
     }
 }
 
-fn prefixes(text: &str) -> impl Iterator<Item = &str> {
-    (0..text.len() - 1).map(|ix| &text[..ix + 1])
-}
-
 fn merge_ranges(ranges: &mut Vec<Range<Anchor>>, buffer: &MultiBufferSnapshot) {
     ranges.sort_unstable_by(|a, b| {
         a.start
@@ -3666,434 +1723,5 @@ fn merge_ranges(ranges: &mut Vec<Range<Anchor>>, buffer: &MultiBufferSnapshot) {
         } else {
             ix += 1;
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use futures::stream::{self};
-    use gpui::{Context, TestAppContext};
-    use indoc::indoc;
-    use language::{
-        language_settings, tree_sitter_rust, Buffer, Language, LanguageConfig, LanguageMatcher,
-        Point,
-    };
-    use language_model::LanguageModelRegistry;
-    use rand::prelude::*;
-    use serde::Serialize;
-    use settings::SettingsStore;
-    use std::{future, sync::Arc};
-
-    #[derive(Serialize)]
-    pub struct DummyCompletionRequest {
-        pub name: String,
-    }
-
-    #[gpui::test(iterations = 10)]
-    async fn test_transform_autoindent(cx: &mut TestAppContext, mut rng: StdRng) {
-        cx.set_global(cx.update(SettingsStore::test));
-        cx.update(language_model::LanguageModelRegistry::test);
-        cx.update(language_settings::init);
-
-        let text = indoc! {"
-            fn main() {
-                let x = 0;
-                for _ in 0..10 {
-                    x += 1;
-                }
-            }
-        "};
-        let buffer =
-            cx.new_model(|cx| Buffer::local(text, cx).with_language(Arc::new(rust_lang()), cx));
-        let buffer = cx.new_model(|cx| MultiBuffer::singleton(buffer, cx));
-        let range = buffer.read_with(cx, |buffer, cx| {
-            let snapshot = buffer.snapshot(cx);
-            snapshot.anchor_before(Point::new(1, 0))..snapshot.anchor_after(Point::new(4, 5))
-        });
-        let prompt_builder = Arc::new(PromptBuilder::new(None).unwrap());
-        let codegen = cx.new_model(|cx| {
-            CodegenAlternative::new(
-                buffer.clone(),
-                range.clone(),
-                true,
-                None,
-                None,
-                prompt_builder,
-                cx,
-            )
-        });
-
-        let chunks_tx = simulate_response_stream(codegen.clone(), cx);
-
-        let mut new_text = concat!(
-            "       let mut x = 0;\n",
-            "       while x < 10 {\n",
-            "           x += 1;\n",
-            "       }",
-        );
-        while !new_text.is_empty() {
-            let max_len = cmp::min(new_text.len(), 10);
-            let len = rng.gen_range(1..=max_len);
-            let (chunk, suffix) = new_text.split_at(len);
-            chunks_tx.unbounded_send(chunk.to_string()).unwrap();
-            new_text = suffix;
-            cx.background_executor.run_until_parked();
-        }
-        drop(chunks_tx);
-        cx.background_executor.run_until_parked();
-
-        assert_eq!(
-            buffer.read_with(cx, |buffer, cx| buffer.snapshot(cx).text()),
-            indoc! {"
-                fn main() {
-                    let mut x = 0;
-                    while x < 10 {
-                        x += 1;
-                    }
-                }
-            "}
-        );
-    }
-
-    #[gpui::test(iterations = 10)]
-    async fn test_autoindent_when_generating_past_indentation(
-        cx: &mut TestAppContext,
-        mut rng: StdRng,
-    ) {
-        cx.set_global(cx.update(SettingsStore::test));
-        cx.update(language_settings::init);
-
-        let text = indoc! {"
-            fn main() {
-                le
-            }
-        "};
-        let buffer =
-            cx.new_model(|cx| Buffer::local(text, cx).with_language(Arc::new(rust_lang()), cx));
-        let buffer = cx.new_model(|cx| MultiBuffer::singleton(buffer, cx));
-        let range = buffer.read_with(cx, |buffer, cx| {
-            let snapshot = buffer.snapshot(cx);
-            snapshot.anchor_before(Point::new(1, 6))..snapshot.anchor_after(Point::new(1, 6))
-        });
-        let prompt_builder = Arc::new(PromptBuilder::new(None).unwrap());
-        let codegen = cx.new_model(|cx| {
-            CodegenAlternative::new(
-                buffer.clone(),
-                range.clone(),
-                true,
-                None,
-                None,
-                prompt_builder,
-                cx,
-            )
-        });
-
-        let chunks_tx = simulate_response_stream(codegen.clone(), cx);
-
-        cx.background_executor.run_until_parked();
-
-        let mut new_text = concat!(
-            "t mut x = 0;\n",
-            "while x < 10 {\n",
-            "    x += 1;\n",
-            "}", //
-        );
-        while !new_text.is_empty() {
-            let max_len = cmp::min(new_text.len(), 10);
-            let len = rng.gen_range(1..=max_len);
-            let (chunk, suffix) = new_text.split_at(len);
-            chunks_tx.unbounded_send(chunk.to_string()).unwrap();
-            new_text = suffix;
-            cx.background_executor.run_until_parked();
-        }
-        drop(chunks_tx);
-        cx.background_executor.run_until_parked();
-
-        assert_eq!(
-            buffer.read_with(cx, |buffer, cx| buffer.snapshot(cx).text()),
-            indoc! {"
-                fn main() {
-                    let mut x = 0;
-                    while x < 10 {
-                        x += 1;
-                    }
-                }
-            "}
-        );
-    }
-
-    #[gpui::test(iterations = 10)]
-    async fn test_autoindent_when_generating_before_indentation(
-        cx: &mut TestAppContext,
-        mut rng: StdRng,
-    ) {
-        cx.update(LanguageModelRegistry::test);
-        cx.set_global(cx.update(SettingsStore::test));
-        cx.update(language_settings::init);
-
-        let text = concat!(
-            "fn main() {\n",
-            "  \n",
-            "}\n" //
-        );
-        let buffer =
-            cx.new_model(|cx| Buffer::local(text, cx).with_language(Arc::new(rust_lang()), cx));
-        let buffer = cx.new_model(|cx| MultiBuffer::singleton(buffer, cx));
-        let range = buffer.read_with(cx, |buffer, cx| {
-            let snapshot = buffer.snapshot(cx);
-            snapshot.anchor_before(Point::new(1, 2))..snapshot.anchor_after(Point::new(1, 2))
-        });
-        let prompt_builder = Arc::new(PromptBuilder::new(None).unwrap());
-        let codegen = cx.new_model(|cx| {
-            CodegenAlternative::new(
-                buffer.clone(),
-                range.clone(),
-                true,
-                None,
-                None,
-                prompt_builder,
-                cx,
-            )
-        });
-
-        let chunks_tx = simulate_response_stream(codegen.clone(), cx);
-
-        cx.background_executor.run_until_parked();
-
-        let mut new_text = concat!(
-            "let mut x = 0;\n",
-            "while x < 10 {\n",
-            "    x += 1;\n",
-            "}", //
-        );
-        while !new_text.is_empty() {
-            let max_len = cmp::min(new_text.len(), 10);
-            let len = rng.gen_range(1..=max_len);
-            let (chunk, suffix) = new_text.split_at(len);
-            chunks_tx.unbounded_send(chunk.to_string()).unwrap();
-            new_text = suffix;
-            cx.background_executor.run_until_parked();
-        }
-        drop(chunks_tx);
-        cx.background_executor.run_until_parked();
-
-        assert_eq!(
-            buffer.read_with(cx, |buffer, cx| buffer.snapshot(cx).text()),
-            indoc! {"
-                fn main() {
-                    let mut x = 0;
-                    while x < 10 {
-                        x += 1;
-                    }
-                }
-            "}
-        );
-    }
-
-    #[gpui::test(iterations = 10)]
-    async fn test_autoindent_respects_tabs_in_selection(cx: &mut TestAppContext) {
-        cx.update(LanguageModelRegistry::test);
-        cx.set_global(cx.update(SettingsStore::test));
-        cx.update(language_settings::init);
-
-        let text = indoc! {"
-            func main() {
-            \tx := 0
-            \tfor i := 0; i < 10; i++ {
-            \t\tx++
-            \t}
-            }
-        "};
-        let buffer = cx.new_model(|cx| Buffer::local(text, cx));
-        let buffer = cx.new_model(|cx| MultiBuffer::singleton(buffer, cx));
-        let range = buffer.read_with(cx, |buffer, cx| {
-            let snapshot = buffer.snapshot(cx);
-            snapshot.anchor_before(Point::new(0, 0))..snapshot.anchor_after(Point::new(4, 2))
-        });
-        let prompt_builder = Arc::new(PromptBuilder::new(None).unwrap());
-        let codegen = cx.new_model(|cx| {
-            CodegenAlternative::new(
-                buffer.clone(),
-                range.clone(),
-                true,
-                None,
-                None,
-                prompt_builder,
-                cx,
-            )
-        });
-
-        let chunks_tx = simulate_response_stream(codegen.clone(), cx);
-        let new_text = concat!(
-            "func main() {\n",
-            "\tx := 0\n",
-            "\tfor x < 10 {\n",
-            "\t\tx++\n",
-            "\t}", //
-        );
-        chunks_tx.unbounded_send(new_text.to_string()).unwrap();
-        drop(chunks_tx);
-        cx.background_executor.run_until_parked();
-
-        assert_eq!(
-            buffer.read_with(cx, |buffer, cx| buffer.snapshot(cx).text()),
-            indoc! {"
-                func main() {
-                \tx := 0
-                \tfor x < 10 {
-                \t\tx++
-                \t}
-                }
-            "}
-        );
-    }
-
-    #[gpui::test]
-    async fn test_inactive_codegen_alternative(cx: &mut TestAppContext) {
-        cx.update(LanguageModelRegistry::test);
-        cx.set_global(cx.update(SettingsStore::test));
-        cx.update(language_settings::init);
-
-        let text = indoc! {"
-            fn main() {
-                let x = 0;
-            }
-        "};
-        let buffer =
-            cx.new_model(|cx| Buffer::local(text, cx).with_language(Arc::new(rust_lang()), cx));
-        let buffer = cx.new_model(|cx| MultiBuffer::singleton(buffer, cx));
-        let range = buffer.read_with(cx, |buffer, cx| {
-            let snapshot = buffer.snapshot(cx);
-            snapshot.anchor_before(Point::new(1, 0))..snapshot.anchor_after(Point::new(1, 14))
-        });
-        let prompt_builder = Arc::new(PromptBuilder::new(None).unwrap());
-        let codegen = cx.new_model(|cx| {
-            CodegenAlternative::new(
-                buffer.clone(),
-                range.clone(),
-                false,
-                None,
-                None,
-                prompt_builder,
-                cx,
-            )
-        });
-
-        let chunks_tx = simulate_response_stream(codegen.clone(), cx);
-        chunks_tx
-            .unbounded_send("let mut x = 0;\nx += 1;".to_string())
-            .unwrap();
-        drop(chunks_tx);
-        cx.run_until_parked();
-
-        // The codegen is inactive, so the buffer doesn't get modified.
-        assert_eq!(
-            buffer.read_with(cx, |buffer, cx| buffer.snapshot(cx).text()),
-            text
-        );
-
-        // Activating the codegen applies the changes.
-        codegen.update(cx, |codegen, cx| codegen.set_active(true, cx));
-        assert_eq!(
-            buffer.read_with(cx, |buffer, cx| buffer.snapshot(cx).text()),
-            indoc! {"
-                fn main() {
-                    let mut x = 0;
-                    x += 1;
-                }
-            "}
-        );
-
-        // Deactivating the codegen undoes the changes.
-        codegen.update(cx, |codegen, cx| codegen.set_active(false, cx));
-        cx.run_until_parked();
-        assert_eq!(
-            buffer.read_with(cx, |buffer, cx| buffer.snapshot(cx).text()),
-            text
-        );
-    }
-
-    #[gpui::test]
-    async fn test_strip_invalid_spans_from_codeblock() {
-        assert_chunks("Lorem ipsum dolor", "Lorem ipsum dolor").await;
-        assert_chunks("```\nLorem ipsum dolor", "Lorem ipsum dolor").await;
-        assert_chunks("```\nLorem ipsum dolor\n```", "Lorem ipsum dolor").await;
-        assert_chunks(
-            "```html\n```js\nLorem ipsum dolor\n```\n```",
-            "```js\nLorem ipsum dolor\n```",
-        )
-        .await;
-        assert_chunks("``\nLorem ipsum dolor\n```", "``\nLorem ipsum dolor\n```").await;
-        assert_chunks("Lorem<|CURSOR|> ipsum", "Lorem ipsum").await;
-        assert_chunks("Lorem ipsum", "Lorem ipsum").await;
-        assert_chunks("```\n<|CURSOR|>Lorem ipsum\n```", "Lorem ipsum").await;
-
-        async fn assert_chunks(text: &str, expected_text: &str) {
-            for chunk_size in 1..=text.len() {
-                let actual_text = StripInvalidSpans::new(chunks(text, chunk_size))
-                    .map(|chunk| chunk.unwrap())
-                    .collect::<String>()
-                    .await;
-                assert_eq!(
-                    actual_text, expected_text,
-                    "failed to strip invalid spans, chunk size: {}",
-                    chunk_size
-                );
-            }
-        }
-
-        fn chunks(text: &str, size: usize) -> impl Stream<Item = Result<String>> {
-            stream::iter(
-                text.chars()
-                    .collect::<Vec<_>>()
-                    .chunks(size)
-                    .map(|chunk| Ok(chunk.iter().collect::<String>()))
-                    .collect::<Vec<_>>(),
-            )
-        }
-    }
-
-    fn simulate_response_stream(
-        codegen: Model<CodegenAlternative>,
-        cx: &mut TestAppContext,
-    ) -> mpsc::UnboundedSender<String> {
-        let (chunks_tx, chunks_rx) = mpsc::unbounded();
-        codegen.update(cx, |codegen, cx| {
-            codegen.handle_stream(
-                String::new(),
-                String::new(),
-                None,
-                future::ready(Ok(LanguageModelTextStream {
-                    message_id: None,
-                    stream: chunks_rx.map(Ok).boxed(),
-                })),
-                cx,
-            );
-        });
-        chunks_tx
-    }
-
-    fn rust_lang() -> Language {
-        Language::new(
-            LanguageConfig {
-                name: "Rust".into(),
-                matcher: LanguageMatcher {
-                    path_suffixes: vec!["rs".to_string()],
-                    ..Default::default()
-                },
-                ..Default::default()
-            },
-            Some(tree_sitter_rust::LANGUAGE.into()),
-        )
-        .with_indents_query(
-            r#"
-            (call_expression) @indent
-            (field_expression) @indent
-            (_ "(" ")" @end) @indent
-            (_ "{" "}" @end) @indent
-            "#,
-        )
-        .unwrap()
     }
 }
