+++ crates/assistant2/src/inline_prompt_editor.rs
@@ -1,5 +1,787 @@
-use gpui::{AnyElement, EventEmitter};
-use ui::{prelude::*, IconButtonShape, Tooltip};
+use crate::buffer_codegen::BufferCodegen;
+use crate::context_picker::ContextPicker;
+use crate::context_store::ContextStore;
+use crate::context_strip::ContextStrip;
+use crate::terminal_codegen::TerminalCodegen;
+use crate::thread_store::ThreadStore;
+use crate::ToggleContextPicker;
+use crate::{
+    assistant_settings::AssistantSettings, CycleNextInlineAssist, CyclePreviousInlineAssist,
+};
+use client::ErrorExt;
+use collections::VecDeque;
+use editor::{
+    actions::{MoveDown, MoveUp},
+    Editor, EditorElement, EditorEvent, EditorMode, EditorStyle, GutterDimensions, MultiBuffer,
+};
+use feature_flags::{FeatureFlagAppExt as _, ZedPro};
+use fs::Fs;
+use gpui::{
+    anchored, deferred, point, AnyElement, AppContext, ClickEvent, CursorStyle, EventEmitter,
+    FocusHandle, FocusableView, FontWeight, Model, Subscription, TextStyle, View, ViewContext,
+    WeakModel, WeakView, WindowContext,
+};
+use language_model::{LanguageModel, LanguageModelRegistry};
+use language_model_selector::{LanguageModelSelector, LanguageModelSelectorPopoverMenu};
+use parking_lot::Mutex;
+use settings::{update_settings_file, Settings};
+use std::sync::Arc;
+use theme::ThemeSettings;
+use ui::{
+    prelude::*, CheckboxWithLabel, IconButtonShape, KeyBinding, Popover, PopoverMenuHandle, Tooltip,
+};
+use util::ResultExt;
+use workspace::Workspace;
+
+#[derive(Copy, Clone, Default, Debug, PartialEq, Eq, Hash)]
+pub struct InlineAssistId(pub usize);
+
+impl InlineAssistId {
+    pub fn post_inc(&mut self) -> InlineAssistId {
+        let id = *self;
+        self.0 += 1;
+        id
+    }
+}
+
+pub struct PromptEditor {
+    pub id: InlineAssistId,
+    pub editor: View<Editor>,
+    // todo (agus): try to make private
+    pub mode: PromptEditorMode,
+    context_strip: View<ContextStrip>,
+    context_picker_menu_handle: PopoverMenuHandle<ContextPicker>,
+    language_model_selector: View<LanguageModelSelector>,
+    edited_since_done: bool,
+    prompt_history: VecDeque<String>,
+    prompt_history_ix: Option<usize>,
+    pending_prompt: String,
+    _codegen_subscription: Subscription,
+    editor_subscriptions: Vec<Subscription>,
+    show_rate_limit_notice: bool,
+}
+
+impl EventEmitter<PromptEditorEvent> for PromptEditor {}
+
+impl Render for PromptEditor {
+    fn render(&mut self, cx: &mut ViewContext<Self>) -> impl IntoElement {
+        let mut buttons = Vec::new();
+
+        // TODO (agus): remove prompt mode and move buttons view here
+        let (prompt_mode, spacing) = match &self.mode {
+            PromptEditorMode::Buffer {
+                codegen,
+                gutter_dimensions,
+            } => {
+                let codegen = codegen.read(cx);
+
+                if codegen.alternative_count(cx) > 1 {
+                    buttons.push(self.render_cycle_controls(cx));
+                }
+
+                let prompt_mode = if codegen.is_insertion {
+                    PromptMode::Generate {
+                        supports_execute: false,
+                    }
+                } else {
+                    PromptMode::Transform
+                };
+
+                let gutter_dimensions = gutter_dimensions.lock();
+
+                let spacing = gutter_dimensions.full_width() + (gutter_dimensions.margin / 2.0);
+
+                (prompt_mode, spacing)
+            }
+            PromptEditorMode::Terminal { codegen: _ } => {
+                // todo(terminal-prompt-cycle)
+                (
+                    PromptMode::Generate {
+                        supports_execute: true,
+                    },
+                    Pixels::ZERO,
+                )
+            }
+        };
+
+        buttons.extend(render_cancel_button(
+            self.codegen_status(cx).into(),
+            self.edited_since_done,
+            prompt_mode,
+            cx,
+        ));
+
+        v_flex()
+            .border_y_1()
+            .border_color(cx.theme().status().info_border)
+            .size_full()
+            .py(cx.line_height() / 2.5)
+            .child(
+                h_flex()
+                    .key_context("PromptEditor")
+                    .bg(cx.theme().colors().editor_background)
+                    .block_mouse_down()
+                    .cursor(CursorStyle::Arrow)
+                    .on_action(cx.listener(Self::toggle_context_picker))
+                    .on_action(cx.listener(Self::confirm))
+                    .on_action(cx.listener(Self::cancel))
+                    .on_action(cx.listener(Self::move_up))
+                    .on_action(cx.listener(Self::move_down))
+                    .capture_action(cx.listener(Self::cycle_prev))
+                    .capture_action(cx.listener(Self::cycle_next))
+                    .child(
+                        h_flex()
+                            .w(spacing)
+                            .justify_center()
+                            .gap_2()
+                            .child(LanguageModelSelectorPopoverMenu::new(
+                                self.language_model_selector.clone(),
+                                IconButton::new("context", IconName::SettingsAlt)
+                                    .shape(IconButtonShape::Square)
+                                    .icon_size(IconSize::Small)
+                                    .icon_color(Color::Muted)
+                                    .tooltip(move |cx| {
+                                        Tooltip::with_meta(
+                                            format!(
+                                                "Using {}",
+                                                LanguageModelRegistry::read_global(cx)
+                                                    .active_model()
+                                                    .map(|model| model.name().0)
+                                                    .unwrap_or_else(|| "No model selected".into()),
+                                            ),
+                                            None,
+                                            "Change Model",
+                                            cx,
+                                        )
+                                    }),
+                            ))
+                            .map(|el| {
+                                let CodegenStatus::Error(error) = self.codegen_status(cx) else {
+                                    return el;
+                                };
+
+                                let error_message = SharedString::from(error.to_string());
+                                if error.error_code() == proto::ErrorCode::RateLimitExceeded
+                                    && cx.has_flag::<ZedPro>()
+                                {
+                                    el.child(
+                                        v_flex()
+                                            .child(
+                                                IconButton::new(
+                                                    "rate-limit-error",
+                                                    IconName::XCircle,
+                                                )
+                                                .toggle_state(self.show_rate_limit_notice)
+                                                .shape(IconButtonShape::Square)
+                                                .icon_size(IconSize::Small)
+                                                .on_click(
+                                                    cx.listener(Self::toggle_rate_limit_notice),
+                                                ),
+                                            )
+                                            .children(self.show_rate_limit_notice.then(|| {
+                                                deferred(
+                                                    anchored()
+                                                        .position_mode(
+                                                            gpui::AnchoredPositionMode::Local,
+                                                        )
+                                                        .position(point(px(0.), px(24.)))
+                                                        .anchor(gpui::Corner::TopLeft)
+                                                        .child(self.render_rate_limit_notice(cx)),
+                                                )
+                                            })),
+                                    )
+                                } else {
+                                    el.child(
+                                        div()
+                                            .id("error")
+                                            .tooltip(move |cx| {
+                                                Tooltip::text(error_message.clone(), cx)
+                                            })
+                                            .child(
+                                                Icon::new(IconName::XCircle)
+                                                    .size(IconSize::Small)
+                                                    .color(Color::Error),
+                                            ),
+                                    )
+                                }
+                            }),
+                    )
+                    .child(div().flex_1().child(self.render_editor(cx)))
+                    .child(h_flex().gap_2().pr_6().children(buttons)),
+            )
+            .child(
+                h_flex()
+                    .child(h_flex().w(spacing).justify_center().gap_2())
+                    .child(self.context_strip.clone()),
+            )
+    }
+}
+
+impl FocusableView for PromptEditor {
+    fn focus_handle(&self, cx: &AppContext) -> FocusHandle {
+        self.editor.focus_handle(cx)
+    }
+}
+
+impl PromptEditor {
+    const MAX_LINES: u8 = 8;
+
+    #[allow(clippy::too_many_arguments)]
+    pub fn new_buffer(
+        id: InlineAssistId,
+        gutter_dimensions: Arc<Mutex<GutterDimensions>>,
+        prompt_history: VecDeque<String>,
+        prompt_buffer: Model<MultiBuffer>,
+        codegen: Model<BufferCodegen>,
+        fs: Arc<dyn Fs>,
+        context_store: Model<ContextStore>,
+        workspace: WeakView<Workspace>,
+        thread_store: Option<WeakModel<ThreadStore>>,
+        cx: &mut ViewContext<Self>,
+    ) -> Self {
+        let codegen_subscription = cx.observe(&codegen, Self::handle_codegen_changed);
+        let mode = PromptEditorMode::Buffer {
+            codegen,
+            gutter_dimensions,
+        };
+
+        let prompt_editor = cx.new_view(|cx| {
+            let mut editor = Editor::new(
+                EditorMode::AutoHeight {
+                    max_lines: Self::MAX_LINES as usize,
+                },
+                prompt_buffer,
+                None,
+                false,
+                cx,
+            );
+            editor.set_soft_wrap_mode(language::language_settings::SoftWrap::EditorWidth, cx);
+            // Since the prompt editors for all inline assistants are linked,
+            // always show the cursor (even when it isn't focused) because
+            // typing in one will make what you typed appear in all of them.
+            editor.set_show_cursor_when_unfocused(true, cx);
+            editor.set_placeholder_text(Self::placeholder_text(&mode, cx), cx);
+            editor
+        });
+        let context_picker_menu_handle = PopoverMenuHandle::default();
+
+        let mut this = Self {
+            id,
+            editor: prompt_editor.clone(),
+            context_strip: cx.new_view(|cx| {
+                ContextStrip::new(
+                    context_store,
+                    workspace.clone(),
+                    thread_store.clone(),
+                    prompt_editor.focus_handle(cx),
+                    context_picker_menu_handle.clone(),
+                    cx,
+                )
+            }),
+            context_picker_menu_handle,
+            language_model_selector: cx.new_view(|cx| {
+                let fs = fs.clone();
+                LanguageModelSelector::new(
+                    move |model, cx| {
+                        update_settings_file::<AssistantSettings>(
+                            fs.clone(),
+                            cx,
+                            move |settings, _| settings.set_model(model.clone()),
+                        );
+                    },
+                    cx,
+                )
+            }),
+            edited_since_done: false,
+            prompt_history,
+            prompt_history_ix: None,
+            pending_prompt: String::new(),
+            _codegen_subscription: codegen_subscription,
+            editor_subscriptions: Vec::new(),
+            show_rate_limit_notice: false,
+            mode,
+        };
+        this.subscribe_to_editor(cx);
+        this
+    }
+
+    fn codegen_status<'a>(&'a self, cx: &'a AppContext) -> &'a CodegenStatus {
+        match &self.mode {
+            PromptEditorMode::Buffer { codegen, .. } => codegen.read(cx).status(cx),
+            PromptEditorMode::Terminal { codegen } => &codegen.read(cx).status,
+        }
+    }
+
+    fn subscribe_to_editor(&mut self, cx: &mut ViewContext<Self>) {
+        self.editor_subscriptions.clear();
+        self.editor_subscriptions
+            .push(cx.subscribe(&self.editor, Self::handle_prompt_editor_events));
+    }
+
+    pub fn set_show_cursor_when_unfocused(
+        &mut self,
+        show_cursor_when_unfocused: bool,
+        cx: &mut ViewContext<Self>,
+    ) {
+        self.editor.update(cx, |editor, cx| {
+            editor.set_show_cursor_when_unfocused(show_cursor_when_unfocused, cx)
+        });
+    }
+
+    pub fn unlink(&mut self, cx: &mut ViewContext<Self>) {
+        let prompt = self.prompt(cx);
+        let focus = self.editor.focus_handle(cx).contains_focused(cx);
+        self.editor = cx.new_view(|cx| {
+            let mut editor = Editor::auto_height(Self::MAX_LINES as usize, cx);
+            editor.set_soft_wrap_mode(language::language_settings::SoftWrap::EditorWidth, cx);
+            editor.set_placeholder_text(Self::placeholder_text(&self.mode, cx), cx);
+            editor.set_placeholder_text("Add a prompt…", cx);
+            editor.set_text(prompt, cx);
+            if focus {
+                editor.focus(cx);
+            }
+            editor
+        });
+        self.subscribe_to_editor(cx);
+    }
+
+    pub fn placeholder_text(mode: &PromptEditorMode, cx: &WindowContext) -> String {
+        let action = match mode {
+            PromptEditorMode::Buffer { codegen, .. } => {
+                if codegen.read(cx).is_insertion {
+                    "Generate"
+                } else {
+                    "Transform"
+                }
+            }
+            PromptEditorMode::Terminal { .. } => "Generate",
+        };
+
+        let assisant_panel_shortcut = ui::text_for_action(&crate::ToggleFocus, cx)
+            .map(|keybinding| format!("{keybinding} to chat ― "))
+            .unwrap_or_default();
+
+        format!("{action}… ({assisant_panel_shortcut}↓↑ for history)")
+    }
+
+    pub fn prompt(&self, cx: &AppContext) -> String {
+        self.editor.read(cx).text(cx)
+    }
+
+    fn toggle_rate_limit_notice(&mut self, _: &ClickEvent, cx: &mut ViewContext<Self>) {
+        self.show_rate_limit_notice = !self.show_rate_limit_notice;
+        if self.show_rate_limit_notice {
+            cx.focus_view(&self.editor);
+        }
+        cx.notify();
+    }
+
+    fn handle_prompt_editor_events(
+        &mut self,
+        _: View<Editor>,
+        event: &EditorEvent,
+        cx: &mut ViewContext<Self>,
+    ) {
+        match event {
+            EditorEvent::Edited { .. } => {
+                if let Some(workspace) = cx.window_handle().downcast::<Workspace>() {
+                    workspace
+                        .update(cx, |workspace, cx| {
+                            let is_via_ssh = workspace
+                                .project()
+                                .update(cx, |project, _| project.is_via_ssh());
+
+                            workspace
+                                .client()
+                                .telemetry()
+                                .log_edit_event("inline assist", is_via_ssh);
+                        })
+                        .log_err();
+                }
+                let prompt = self.editor.read(cx).text(cx);
+                if self
+                    .prompt_history_ix
+                    .map_or(true, |ix| self.prompt_history[ix] != prompt)
+                {
+                    self.prompt_history_ix.take();
+                    self.pending_prompt = prompt;
+                }
+
+                self.edited_since_done = true;
+                cx.notify();
+            }
+            EditorEvent::Blurred => {
+                if self.show_rate_limit_notice {
+                    self.show_rate_limit_notice = false;
+                    cx.notify();
+                }
+            }
+            _ => {}
+        }
+    }
+
+    fn handle_codegen_changed(&mut self, _: Model<BufferCodegen>, cx: &mut ViewContext<Self>) {
+        match self.codegen_status(cx) {
+            CodegenStatus::Idle => {
+                self.editor
+                    .update(cx, |editor, _| editor.set_read_only(false));
+            }
+            CodegenStatus::Pending => {
+                self.editor
+                    .update(cx, |editor, _| editor.set_read_only(true));
+            }
+            CodegenStatus::Done => {
+                self.edited_since_done = false;
+                self.editor
+                    .update(cx, |editor, _| editor.set_read_only(false));
+            }
+            CodegenStatus::Error(error) => {
+                if cx.has_flag::<ZedPro>()
+                    && error.error_code() == proto::ErrorCode::RateLimitExceeded
+                    && !dismissed_rate_limit_notice()
+                {
+                    self.show_rate_limit_notice = true;
+                    cx.notify();
+                }
+
+                self.edited_since_done = false;
+                self.editor
+                    .update(cx, |editor, _| editor.set_read_only(false));
+            }
+        }
+    }
+
+    fn toggle_context_picker(&mut self, _: &ToggleContextPicker, cx: &mut ViewContext<Self>) {
+        self.context_picker_menu_handle.toggle(cx);
+    }
+
+    fn cancel(&mut self, _: &editor::actions::Cancel, cx: &mut ViewContext<Self>) {
+        match self.codegen_status(cx) {
+            CodegenStatus::Idle | CodegenStatus::Done | CodegenStatus::Error(_) => {
+                cx.emit(PromptEditorEvent::CancelRequested);
+            }
+            CodegenStatus::Pending => {
+                cx.emit(PromptEditorEvent::StopRequested);
+            }
+        }
+    }
+
+    fn confirm(&mut self, _: &menu::Confirm, cx: &mut ViewContext<Self>) {
+        match self.codegen_status(cx) {
+            CodegenStatus::Idle => {
+                cx.emit(PromptEditorEvent::StartRequested);
+            }
+            CodegenStatus::Pending => {
+                cx.emit(PromptEditorEvent::DismissRequested);
+            }
+            CodegenStatus::Done => {
+                if self.edited_since_done {
+                    cx.emit(PromptEditorEvent::StartRequested);
+                } else {
+                    cx.emit(PromptEditorEvent::ConfirmRequested { execute: false });
+                }
+            }
+            CodegenStatus::Error(_) => {
+                cx.emit(PromptEditorEvent::StartRequested);
+            }
+        }
+    }
+
+    fn move_up(&mut self, _: &MoveUp, cx: &mut ViewContext<Self>) {
+        if let Some(ix) = self.prompt_history_ix {
+            if ix > 0 {
+                self.prompt_history_ix = Some(ix - 1);
+                let prompt = self.prompt_history[ix - 1].as_str();
+                self.editor.update(cx, |editor, cx| {
+                    editor.set_text(prompt, cx);
+                    editor.move_to_beginning(&Default::default(), cx);
+                });
+            }
+        } else if !self.prompt_history.is_empty() {
+            self.prompt_history_ix = Some(self.prompt_history.len() - 1);
+            let prompt = self.prompt_history[self.prompt_history.len() - 1].as_str();
+            self.editor.update(cx, |editor, cx| {
+                editor.set_text(prompt, cx);
+                editor.move_to_beginning(&Default::default(), cx);
+            });
+        }
+    }
+
+    fn move_down(&mut self, _: &MoveDown, cx: &mut ViewContext<Self>) {
+        if let Some(ix) = self.prompt_history_ix {
+            if ix < self.prompt_history.len() - 1 {
+                self.prompt_history_ix = Some(ix + 1);
+                let prompt = self.prompt_history[ix + 1].as_str();
+                self.editor.update(cx, |editor, cx| {
+                    editor.set_text(prompt, cx);
+                    editor.move_to_end(&Default::default(), cx)
+                });
+            } else {
+                self.prompt_history_ix = None;
+                let prompt = self.pending_prompt.as_str();
+                self.editor.update(cx, |editor, cx| {
+                    editor.set_text(prompt, cx);
+                    editor.move_to_end(&Default::default(), cx)
+                });
+            }
+        }
+    }
+
+    fn cycle_prev(&mut self, _: &CyclePreviousInlineAssist, cx: &mut ViewContext<Self>) {
+        match &self.mode {
+            PromptEditorMode::Buffer { codegen, .. } => {
+                codegen.update(cx, |codegen, cx| codegen.cycle_prev(cx));
+            }
+            PromptEditorMode::Terminal { codegen } => todo!("terminal-prompt-cycle"),
+        }
+    }
+
+    fn cycle_next(&mut self, _: &CycleNextInlineAssist, cx: &mut ViewContext<Self>) {
+        match &self.mode {
+            PromptEditorMode::Buffer { codegen, .. } => {
+                codegen.update(cx, |codegen, cx| codegen.cycle_prev(cx));
+            }
+            PromptEditorMode::Terminal { codegen } => todo!("terminal-prompt-cycle"),
+        }
+    }
+
+    fn render_cycle_controls(&self, cx: &ViewContext<Self>) -> AnyElement {
+        let PromptEditorMode::Buffer { codegen, .. } = &self.mode else {
+            // todo(terminal-prompt-cycle)
+            return div().into_any_element();
+        };
+
+        let codegen = codegen.read(cx);
+        let disabled = matches!(codegen.status(cx), CodegenStatus::Idle);
+
+        let model_registry = LanguageModelRegistry::read_global(cx);
+        let default_model = model_registry.active_model();
+        let alternative_models = model_registry.inline_alternative_models();
+
+        let get_model_name = |index: usize| -> String {
+            let name = |model: &Arc<dyn LanguageModel>| model.name().0.to_string();
+
+            match index {
+                0 => default_model.as_ref().map_or_else(String::new, name),
+                index if index <= alternative_models.len() => alternative_models
+                    .get(index - 1)
+                    .map_or_else(String::new, name),
+                _ => String::new(),
+            }
+        };
+
+        let total_models = alternative_models.len() + 1;
+
+        if total_models <= 1 {
+            return div().into_any_element();
+        }
+
+        let current_index = codegen.active_alternative;
+        let prev_index = (current_index + total_models - 1) % total_models;
+        let next_index = (current_index + 1) % total_models;
+
+        let prev_model_name = get_model_name(prev_index);
+        let next_model_name = get_model_name(next_index);
+
+        h_flex()
+            .child(
+                IconButton::new("previous", IconName::ChevronLeft)
+                    .icon_color(Color::Muted)
+                    .disabled(disabled || current_index == 0)
+                    .shape(IconButtonShape::Square)
+                    .tooltip({
+                        let focus_handle = self.editor.focus_handle(cx);
+                        move |cx| {
+                            cx.new_view(|cx| {
+                                let mut tooltip = Tooltip::new("Previous Alternative").key_binding(
+                                    KeyBinding::for_action_in(
+                                        &CyclePreviousInlineAssist,
+                                        &focus_handle,
+                                        cx,
+                                    ),
+                                );
+                                if !disabled && current_index != 0 {
+                                    tooltip = tooltip.meta(prev_model_name.clone());
+                                }
+                                tooltip
+                            })
+                            .into()
+                        }
+                    })
+                    .on_click(cx.listener(|this, _, cx| {
+                        // this.codegen
+                        //     .update(cx, |codegen, cx| codegen.cycle_prev(cx))
+                        todo!("terminal-prompt-cycle")
+                    })),
+            )
+            .child(
+                Label::new(format!(
+                    "{}/{}",
+                    codegen.active_alternative + 1,
+                    codegen.alternative_count(cx)
+                ))
+                .size(LabelSize::Small)
+                .color(if disabled {
+                    Color::Disabled
+                } else {
+                    Color::Muted
+                }),
+            )
+            .child(
+                IconButton::new("next", IconName::ChevronRight)
+                    .icon_color(Color::Muted)
+                    .disabled(disabled || current_index == total_models - 1)
+                    .shape(IconButtonShape::Square)
+                    .tooltip({
+                        let focus_handle = self.editor.focus_handle(cx);
+                        move |cx| {
+                            cx.new_view(|cx| {
+                                let mut tooltip = Tooltip::new("Next Alternative").key_binding(
+                                    KeyBinding::for_action_in(
+                                        &CycleNextInlineAssist,
+                                        &focus_handle,
+                                        cx,
+                                    ),
+                                );
+                                if !disabled && current_index != total_models - 1 {
+                                    tooltip = tooltip.meta(next_model_name.clone());
+                                }
+                                tooltip
+                            })
+                            .into()
+                        }
+                    })
+                    .on_click(cx.listener(|this, _, cx| {
+                        // this.codegen
+                        //     .update(cx, |codegen, cx| codegen.cycle_next(cx))
+                        todo!("terminal-prompt-cycle")
+                    })),
+            )
+            .into_any_element()
+    }
+
+    fn render_rate_limit_notice(&self, cx: &mut ViewContext<Self>) -> impl IntoElement {
+        Popover::new().child(
+            v_flex()
+                .occlude()
+                .p_2()
+                .child(
+                    Label::new("Out of Tokens")
+                        .size(LabelSize::Small)
+                        .weight(FontWeight::BOLD),
+                )
+                .child(Label::new(
+                    "Try Zed Pro for higher limits, a wider range of models, and more.",
+                ))
+                .child(
+                    h_flex()
+                        .justify_between()
+                        .child(CheckboxWithLabel::new(
+                            "dont-show-again",
+                            Label::new("Don't show again"),
+                            if dismissed_rate_limit_notice() {
+                                ui::ToggleState::Selected
+                            } else {
+                                ui::ToggleState::Unselected
+                            },
+                            |selection, cx| {
+                                let is_dismissed = match selection {
+                                    ui::ToggleState::Unselected => false,
+                                    ui::ToggleState::Indeterminate => return,
+                                    ui::ToggleState::Selected => true,
+                                };
+
+                                set_rate_limit_notice_dismissed(is_dismissed, cx)
+                            },
+                        ))
+                        .child(
+                            h_flex()
+                                .gap_2()
+                                .child(
+                                    Button::new("dismiss", "Dismiss")
+                                        .style(ButtonStyle::Transparent)
+                                        .on_click(cx.listener(Self::toggle_rate_limit_notice)),
+                                )
+                                .child(Button::new("more-info", "More Info").on_click(
+                                    |_event, cx| {
+                                        cx.dispatch_action(Box::new(
+                                            zed_actions::OpenAccountSettings,
+                                        ))
+                                    },
+                                )),
+                        ),
+                ),
+        )
+    }
+
+    fn render_editor(&mut self, cx: &mut ViewContext<Self>) -> AnyElement {
+        let font_size = TextSize::Default.rems(cx);
+        let line_height = font_size.to_pixels(cx.rem_size()) * 1.3;
+
+        v_flex()
+            .key_context("MessageEditor")
+            .size_full()
+            .gap_2()
+            .p_2()
+            .bg(cx.theme().colors().editor_background)
+            .child({
+                let settings = ThemeSettings::get_global(cx);
+                let text_style = TextStyle {
+                    color: cx.theme().colors().editor_foreground,
+                    font_family: settings.ui_font.family.clone(),
+                    font_features: settings.ui_font.features.clone(),
+                    font_size: font_size.into(),
+                    font_weight: settings.ui_font.weight,
+                    line_height: line_height.into(),
+                    ..Default::default()
+                };
+
+                EditorElement::new(
+                    &self.editor,
+                    EditorStyle {
+                        background: cx.theme().colors().editor_background,
+                        local_player: cx.theme().players().local(),
+                        text: text_style,
+                        ..Default::default()
+                    },
+                )
+            })
+            .into_any_element()
+    }
+}
+
+const DISMISSED_RATE_LIMIT_NOTICE_KEY: &str = "dismissed-rate-limit-notice";
+
+fn dismissed_rate_limit_notice() -> bool {
+    db::kvp::KEY_VALUE_STORE
+        .read_kvp(DISMISSED_RATE_LIMIT_NOTICE_KEY)
+        .log_err()
+        .map_or(false, |s| s.is_some())
+}
+
+fn set_rate_limit_notice_dismissed(is_dismissed: bool, cx: &mut AppContext) {
+    db::write_and_log(cx, move || async move {
+        if is_dismissed {
+            db::kvp::KEY_VALUE_STORE
+                .write_kvp(DISMISSED_RATE_LIMIT_NOTICE_KEY.into(), "1".into())
+                .await
+        } else {
+            db::kvp::KEY_VALUE_STORE
+                .delete_kvp(DISMISSED_RATE_LIMIT_NOTICE_KEY.into())
+                .await
+        }
+    })
+}
+
+pub enum PromptEditorMode {
+    Buffer {
+        codegen: Model<BufferCodegen>,
+        gutter_dimensions: Arc<Mutex<GutterDimensions>>,
+    },
+    Terminal {
+        codegen: Model<TerminalCodegen>,
+    },
+}
 
 pub enum CodegenStatus {
     Idle,
@@ -188,4 +970,10 @@ pub enum PromptEditorEvent {
     CancelRequested,
     DismissRequested,
     Resized { height_in_lines: u8 },
+}
+
+#[derive(Copy, Clone, Debug)]
+pub enum CodegenEvent {
+    Finished,
+    Undone,
 }
